### 2025-03-03 #7867 @editorialbot
[PRE REVIEW]: ParquetDB: A Lightweight Python Parquet-Based Database
**Submitting author:** @lllangWV (<a href="http://orcid.org/0000-0003-2867-1706">Logan Lang</a>)
**Repository:** https://github.com/lllangWV/ParquetDB
**Branch with paper.md** (empty if default branch): main
**Version:** v0.25.1
