### 2025-03-03 #7867 @danielskatz
👋 @lllangWV - thanks for your submission. Please add the country to the second affiliation, and note that we don't require addresses for affiliations, just organizations and their location (which could be city and country or just country).

Also, your paper is a bit long, though I'm not sure what I would suggest removing. Please do, however, consider if there's anything that could be in the repo or docs, and just pointed to from the paper, rather than being in the paper.  Perhaps one of the two benchmarks?

Finally, please look at the your references in the paper, and make sure they follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html): there should be a space before a reference, and multiple references have a particular format.

You can use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. And the command `@editorialbot check repository` runs a bunch of checks, including the wordcount of the paper, which should be close to 1000 words.  editorialbot commands need to be the first thing in a new comment.
