### 2025-03-03 a905a78
Update JOSS paper with refined bibliography, streamlined content, and improved references
