+++ joss/paper.md
@@ -24,11 +24,11 @@ authors:
     affiliation: '1'
 affiliations:
   - index: 1
-    name: Department of Physics, West Virginia University, Morgantown, WV 26506, United States
+    name: Department of Physics, West Virginia University, Morgantown, United States
   - index: 2
-    name: Instituto de Ciencia de Materiales de Madrid, Campus de Cantoblanco, C. Sor Juana Inés de la Cruz, 3, Fuencarral-El Pardo, Madrid 28049, Spain
+    name: Instituto de Ciencia de Materiales de Madrid, Madrid, Spain
   - index: 3
-    name: National Institute of Standards and Technology, 100 Bureau Dr, Gaithersburg, MD 20899, United States
+    name: National Institute of Standards and Technology, Gaithersburg, United States
 date: 14 February 2025
 bibliography: ParquetDB.bib
 ---
@@ -39,48 +39,22 @@ ParquetDB is a Python library that serves as a "medium ware" solution, bridging 
 
 # Statement of need
 
-In an era where data is the driving force behind innovation, the demand for highly efficient, scalable, and adaptable storage solutions has never been greater. Traditional file-based storage formats (e.g., CSV, JSON, TXT) and database systems (e.g., SQLite[@SQLite], MongoDB[@MongoDBDeveloperData]) have historically powered data handling in numerous applications [@habyarimanaGenomicsData2021] [@jainCommentaryMaterialsProject2013] [@WellKnownUsersSQLite] [@CustomerCaseStudies]. However, both approaches exhibit unique limitations that can impede rapid experimentation, large-scale research, and data-intensive development.
+In an era where data is the driving force behind innovation, the demand for highly efficient, scalable, and adaptable storage solutions has never been greater. Traditional file-based storage formats (e.g., CSV, JSON, TXT) and database systems (e.g., SQLite [@allenDefinitiveGuideSQLite2010], MongoDB [@guoMongoDBsJavaScriptFuzzer2017]) have historically powered data handling in numerous applications [@habyarimanaGenomicsData2021] [@jainCommentaryMaterialsProject2013] [@hjorthlarsenAtomicSimulationEnvironment2017]. However, both approaches exhibit unique limitations that can impede rapid experimentation, large-scale research, and data-intensive development.
 
 File-based solutions are popular for their simplicity and portability, often relying on straightforward ASCII/UTF encoding. This design choice, while human-readable, becomes highly inefficient for numerical data. For example, encoding an integer like 127 in ASCII demands three separate bytes (00110001 00110010 00110111), leading to significant overhead as data volumes expand. Such inflation in file size translates into slower input/output (I/O) operations and increased storage requirements, ultimately restricting scalability. Additionally, file-based formats typically lack built-in querying capabilities and indexing features, forcing developers to manage complex data relationships manually. These constraints limit the agility of workflows, especially as projects grow in complexity or require quick iteration cycles.
 
-Conversely, traditional database management systems offer robust encoding, indexing, and querying capabilities out of the box. Relational databases, for instance, enforce structured schemas that ensure data integrity but introduce complexities when the data model evolves over time. Non-relational databases, such as document-oriented or key-value stores, are more flexible but risk data inconsistency and can become cumbersome to optimize for performance[@singh15TypesDatabases2024]. Many of these solutions require dedicated servers or intricate configurations, increasing overhead for lightweight experimentation. Moreover, the underlying architectures—whether row-based or reliant on specialized storage engines—can exhibit performance bottlenecks when handling unstructured or semi-structured data at scale.
[…]
