+++ joss/ParquetDB.bib
@@ -1,43 +1,30 @@
-@misc{singh15TypesDatabases2024,
-  title = {15 {{Types}} of {{Databases}} and {{When}} to {{Use Them}}},
-  author = {Singh, Ashish Pratap},
-  year = {2024},
-  month = jul,
-  urldate = {2024-10-20},
-  abstract = {\#4 System Design - Types of Databases},
-  howpublished = {https://blog.algomaster.io/p/15-types-of-databases},
-  langid = {english}
+@book{allenDefinitiveGuideSQLite2010,
+  title = {The {{Definitive Guide}} to {{SQLite}}},
+  author = {Allen, Grant and Owens, Mike},
+  year = {2010},
+  publisher = {Apress},
+  address = {Berkeley, CA},
+  doi = {10.1007/978-1-4302-3226-1},
+  urldate = {2025-03-03},
+  copyright = {http://www.springer.com/tdm},
+  isbn = {978-1-4302-3225-4 978-1-4302-3226-1},
+  langid = {english},
+  keywords = {database,database management,relational database,SQL,SQLite}
 }
 
-@misc{CustomerCaseStudies,
-  title = {Customer {{Case Studies}}},
-  journal = {MongoDB},
-  urldate = {2024-10-20},
-  abstract = {See the leading organizations creating applications never before possible with MongoDB.},
-  howpublished = {https://www.mongodb.com/solutions/customer-case-studies},
-  langid = {english}
-}
-
-@misc{WellKnownUsersSQLite,
-  title = {Well-{{Known Users Of SQLite}}},
-  urldate = {2024-10-20},
-  howpublished = {https://www.sqlite.org/famous.html}
-}
-
-@article{jainCommentaryMaterialsProject2013,
-  title = {Commentary: {{The Materials Project}}: {{A}} Materials Genome Approach to Accelerating Materials Innovation},
-  shorttitle = {Commentary},
-  author = {Jain, Anubhav and Ong, Shyue Ping and Hautier, Geoffroy and Chen, Wei and Richards, William Davidson and Dacek, Stephen and Cholia, Shreyas and Gunter, Dan and Skinner, David and Ceder, Gerbrand and Persson, Kristin A.},
-  year = {2013},
-  month = jul,
-  journal = {APL Materials},
-  volume = {1},
-  number = {1},
-  pages = {011002},
-  issn = {2166-532X},
-  doi = {10.1063/1.4812323},
-  urldate = {2024-10-20},
-  abstract = {Accelerating the discovery of advanced materials is essential for human welfare and sustainable, clean energy. In this paper, we introduce the Materials Project (www.materialsproject.org), a core program of the Materials Genome Initiative that uses high-throughput computing to uncover the properties of all known inorganic materials. This open dataset can be accessed through multiple channels for both interactive exploration and data mining. The Materials Project also seeks to create open-source platforms for developing robust, sophisticated materials analyses. Future efforts will enable users to perform ``rapid-prototyping'' of new materials in silico, and provide researchers with new avenues for cost-effective, data-driven materials design.}
+@article{guoMongoDBsJavaScriptFuzzer2017,
+  title = {{{MongoDB}}'s {{JavaScript}} Fuzzer},
+  author = {Guo, Robert},
+  year = {2017},
+  month = apr,
+  journal = {Commun. ACM},
+  volume = {60},
+  number = {5},
+  pages = {43--47},
+  issn = {0001-0782},
+  doi = {10.1145/3052937},
+  urldate = {2025-03-03},
+  abstract = {The fuzzer is for those edge cases that your testing did not catch.}
 }
 
[…]
