### 2025-03-03 5d32e81
Update JOSS paper citation formatting in bibliography
