+++ joss/paper.md
@@ -39,7 +39,7 @@ ParquetDB is a Python library that serves as a "medium ware" solution, bridging 
 
 # Statement of need
 
-In an era where data is the driving force behind innovation, the demand for highly efficient, scalable, and adaptable storage solutions has never been greater. Traditional file-based storage formats (e.g., CSV, JSON, TXT) and database systems (e.g., SQLite [@allenDefinitiveGuideSQLite2010], MongoDB [@guoMongoDBsJavaScriptFuzzer2017]) have historically powered data handling in numerous applications [@habyarimanaGenomicsData2021] [@jainCommentaryMaterialsProject2013] [@hjorthlarsenAtomicSimulationEnvironment2017]. However, both approaches exhibit unique limitations that can impede rapid experimentation, large-scale research, and data-intensive development.
+In an era where data is the driving force behind innovation, the demand for highly efficient, scalable, and adaptable storage solutions has never been greater. Traditional file-based storage formats (e.g., CSV, JSON, TXT) and database systems (e.g., SQLite [@allenDefinitiveGuideSQLite2010], MongoDB [@guoMongoDBsJavaScriptFuzzer2017]) have historically powered data handling in numerous applications [@habyarimanaGenomicsData2021;@jainCommentaryMaterialsProject2013;@hjorthlarsenAtomicSimulationEnvironment2017]. However, both approaches exhibit unique limitations that can impede rapid experimentation, large-scale research, and data-intensive development.
 
 File-based solutions are popular for their simplicity and portability, often relying on straightforward ASCII/UTF encoding. This design choice, while human-readable, becomes highly inefficient for numerical data. For example, encoding an integer like 127 in ASCII demands three separate bytes (00110001 00110010 00110111), leading to significant overhead as data volumes expand. Such inflation in file size translates into slower input/output (I/O) operations and increased storage requirements, ultimately restricting scalability. Additionally, file-based formats typically lack built-in querying capabilities and indexing features, forcing developers to manage complex data relationships manually. These constraints limit the agility of workflows, especially as projects grow in complexity or require quick iteration cycles.
 
