### 2025-03-03 #7867 @lllangWV
Hey @danielskatz,

Thanks for the feedback! I made the suggested adjustments. I believe I fixed the references now.

As for reducing the length of the paper, I reduced the size to 1076, do you think this size will be good? I can try to reduce it more if I have to. For instance I could remove the installation section.
