### 2025-03-17 #7867 @fabian-s
hi @ckoerber @perdelt

would any of you be willing to review this submission (https://github.com/lllangWV/ParquetDB) for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Even if you are not interested or don't have the time, please reply to this message so we can start looking for other reviewers.
