### 2025-03-21 #7867 @fabian-s
hi @jcrobak @perdelt @wesm

would any of you be willing to review this submission (https://github.com/lllangWV/ParquetDB) for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Even if you are not interested or don't have the time, please be so kind and reply to this message so I can start looking for other reviewers.
