### 2025-03-21 #7932 @editorialbot
[REVIEW]: ParquetDB: A Lightweight Python Parquet-Based Database
**Submitting author:** @lllangWV (<a href="http://orcid.org/0000-0003-2867-1706">Logan Lang</a>)
**Repository:** https://github.com/lllangWV/ParquetDB
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.1
