### 2025-03-31 #7932 @fabian-s
@ckoerber @perdelt please don't forget to `@editorialbot generate my checklist` -- if you think you'll need more than 3 weeks from today for this review, please let me know so I can try to keep things on track.
