### 2025-04-13 lllangWV/ParquetDB#8 @ckoerber
Test Failure & Deprecation Warning with Python 3.10 / ParquetDB 0.25.2.dev18 / pyarrow 19
When running the test suite with ParquetDB 0.25.2.dev18 (Python 3.10.12), I encountered the following test failure:

```
========================================================================== FAILURES ==========================================================================
________________________________________________________ TestPythonObjectTypes.test_arrow_conversion _________________________________________________________

self = <test_types.TestPythonObjectTypes testMethod=test_arrow_conversion>

    def test_arrow_conversion(self):
        # Test conversion from PythonObjectPandasArray to Arrow Array
       for key in self.test_data:
            arr = PythonObjectPandasArray(self.test_data[key])
            arrow_arr = pa.array(arr, type=PythonObjectArrowType())

>           self.assertEqual(arrow_arr.type, PythonObjectArrowType())
E           AssertionError: DataType(binary) != PythonObjectArrowType(DataType(binary))

tests/test_types.py:89: AssertionError
```

In addition, I see a deprecation warning:
```
tests/test_types.py::TestPythonObjectTypes::test_dataframe_integration
  python3.10/site-packages/pandas/core/arrays/base.py:725: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and 'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
    return np.array(self, dtype=dtype, copy=copy)
```

My environment uses (Python 3.10.12):

```
arrow                     1.3.0
numpy                     2.2.4
pyarrow                   19.0.1
pandas                    2.2.3
```

Pleas let me know, if you need more information.

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7932
