### 2025-04-13 lllangWV/ParquetDB#9 @ckoerber
Deprecated ParquetDB Initialization in Benchmarks and Paper
I noticed that some benchmark scripts (and also the Arxiv paper) still use an older version of the _ParquetDB_ initialization with `dataset_name: Optional[str]` and `dir: Optional[str]` arguments (the repo search of `ParquetDB(dataset_name`  has 11 hits).

In the current version of the code, the constructor appears to be:

```python

def __init__(
        self,
        db_path,
        initial_fields: List[pa.Field] = None,
        serialize_python_objects: bool = True,
        use_multiprocessing: bool = False,
    )
```
It might be helpful to update the benchmarks and documentation accordingly to avoid confusion.

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7932
