### 2025-04-13 lllangWV/ParquetDB#10 @ckoerber
Overwriting of Existing Files After Manual Deletion
While testing internal behavior, I noticed that manually deleting one of the internal `.parquet` files (other than the latest one) may lead to data loss.

Example (after creating a few files in the `db_path`):
```
db_path/
    db_0.parquet
    db_1.parquet
    db_2.parquet
```
After deleting `db_1.parquet` (`rm db_path/db_1.parquet`), the next create call will reuse `db_2.parquet` as the filename, overwriting the existing file, since [`_get_save_path`](https://github.com/lllangWV/ParquetDB/blob/dcf32a48fafdce1be40930d19123ca88d42abc9b/parquetdb/core/parquetdb.py#L2848) determines the new filename based on the current file count.

While this is arguably "unexpected user behavior," it may be worth:

* Explicitly warning users in the documentation not to manually modify files in the database directory.
* Alternatively, raising an exception if existing files would be overwritten / the internal logic is broken
* Or using randomized or UUID-based filenames (e.g., via [tempfile with fixed dir](https://docs.python.org/3/library/tempfile.html#tempfile.TemporaryFile)) to prevent accidental overwrites.

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7932
