### 2025-04-13 #7932 @ckoerber
Dear @fabian-s, dear @lllangWV,

I have completed my review of _ParquetDB_. Following the JOSS review guidelines and checklist, I conclude that most of the criteria for publication are fulfilled.

I have opened a few minor issues directly in the repository. These concern small improvements but do not affect the core functionality of _ParquetDB_. I will also summarize a few broader points here for discussion. If preferred, I am happy to open dedicated GitHub issues for these as well.

## Substantial Scholarly Effort

The main point I would like to raise relates to the JOSS criterion of substantial scholarly effort, specifically regarding the objective:

> Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).

The JOSS paper draft does not explicitly target a specific scientific field — which is reasonable, given that managing large datasets is a cross-domain problem. Therefore, I will focus on assessing the contribution of _ParquetDB_ in comparison to existing tools in the general data storage ecosystem.

### Positioning within the Ecosystem

To my understanding, _ParquetDB_ builds upon the Apache Parquet format and the Apache Arrow API. It provides a Python-based interface that enables the features of the Apache stack:

1. Efficient, compressed, and schema-based file storage.
2. Querying capabilities, particularly leveraging batch operations to overcome hardware memory limits.

The key additional contributions of _ParquetDB_ -- compared to the Apache stack -- seem to be:

1. Automatic indexing of stored data objects
2. Data normalization functionality to merge, check, and clean multiple Parquet files into consistent datasets, going beyond basic (immutable) [Parquet Datasets](https://arrow.apache.org/docs/python/parquet.html#partitioned-datasets-multiple-files) features provided by Apache Arrow.

This position of the authors are that _ParquetDB_ is as a lightweight middleware layer between file-based storage and full database systems. While I agree with this positioning, I would argue that _ParquetDB_ is still conceptually closer to file-based storage systems, as typical database features (relations and their integrity) are not implemented.

In this context, a comparison with alternative file-based formats, such as the mentioned MongoDB, Parquets via PyArrow directly and SQLite makes sense. Also, which is not mentioned in the paper, a comparison with [HDF5](https://www.hdfgroup.org/solutions/hdf5/), might also be relevant; particularly when handling array data.

### Assessment of Novelty and Usefulness

While I agree that the setup for a _ParquetDB_ is simple, setting up Parquet files directly (via PyArrow) or using lightweight file-based databases (such as SQLite, which is natively supported in Python and supports JSON data) is not significantly more complicated for many use cases.

The main advantage of _ParquetDB_ over SQLite seems to be the performance for returning large-scale data (e.g., queries returning 100k rows with 100 columns). Note that the filtering of many rows (see benchmark below) has comparable performance. As such, depending on the use case one might argue that filtering and returning smaller datasets within SQLite may alleviate the performance advantage of _ParquetDB_. For example SQL [Window](https://sqlite.org/windowfunctions.html) and [Aggregate functions](https://sqlite.org/lang_aggfunc.html) may be utilized to return smaller datasets and thus provide comparable performance. Hence, the performance benefit of _ParquetDB_ seems to heavily depend on the kind of problem one intends to solve.

Moreover, outside of the normalization functionality, adding an index to a collection of Parquet files could be achieved with relatively little additional code using PyArrow directly.

Therefore, to strengthen the case for publication, I would encourage the authors to elaborate more clearly on:

* In which concrete scenarios _ParquetDB_ provides significant advantages over directly using the Apache stack with some few lines of project specific code ("What functionality does _ParquetDB_ add on top of the existing Dataset functionality that makes live more easy?")?
* In which concrete scenarios does _ParquetDB_ provide significant advantages over directly using relational databases?

To my opinion, a feature that could provide significant added value (but is currently not supported, as noted in the [longer paper on arxiv](https://arxiv.org/abs/2502.05311)) would be concurrent writing — e.g., along the lines of [HDF5 and MPI parallelization](https://support.hdfgroup.org/documentation/hdf5/latest/_intro_par_h_d_f5.html) or distributed writes followed by normalization. Given ParquetDB's design of "create many files first, normalize later," I believe this could be an interesting future direction.

## Opened issues

* [Test Failure & Deprecation Warning with Python 3.10 / ParquetDB 0.25.2.dev18 / pyarrow 19](https://github.com/lllangWV/ParquetDB/issues/8)
* [Deprecated ParquetDB Initialization in Benchmarks and Paper](https://github.com/lllangWV/ParquetDB/issues/9)
* [Overwriting of Existing Files After Manual Deletion](https://github.com/lllangWV/ParquetDB/issues/10)

## Further remarks

### Tests & Coverage

- Automated tests are currently not integrated. On my local system, one of the tests fails (see issue).
- Using [pytest](https://docs.pytest.org/en/stable/index.html) and [coverage](https://coverage.readthedocs.io/en/7.8.0/), I measured a test coverage of approximately 73% overall, and 65% for the most relevant module `parquetdb/core/parquetdb.py`. This seems reasonable but could be improved, particularly for the core functionality (of course the coverage is just a limited predictor for usefulness).

### Citations & References

It may be appropriate to directly cite the following resources in the JOSS paper:

- Apache Software Foundation (Parquet and Arrow projects).
- The longer paper associated with *ParquetDB*: https://arxiv.org/abs/2502.05311.

### Performance Evaluation

- I was unable to execute the provided benchmarks directly (see issue).
- Benchmark results should ideally include the hardware specifications (which are provided in the longer paper, but could be summarized here).
- Based on my own experiments, I find that *ParquetDB* and SQLite perform similarly for selective "needle-in-a-haystack" queries. The performance advantage of *ParquetDB* appears primarily when reading and returning large datasets, rather than during the filtering step.
- The naive implementation of what *ParquetDB* does (using a few lines of PyArrow) outperforms *ParquetDB* in the read/query/write benchmarks

### Miscellaneous Suggestions

- The [detailed contributing file](https://lllangwv.github.io/ParquetDB/CONTRIBUTING.html) is not visible in the repository root directory
- The `requirements.txt` file may no longer be necessary.
- Anaconda environments are listed but not mentioned in the documentation — it might be helpful to clarify this (unless intentionally omitted for licensing reasons).
- There are GitHub Actions configured for documentation generation and PyPI deployment. I also noticed an interesting feature: an [automated changelog generation script using an OpenAI interface](https://github.com/lllangWV/_ParquetDB_/blob/main/scripts/github_workflow/workflow_generate_changlog_on_recent_tag.py). While this is innovative, I would personally prefer that changelog generation remains a manual (or at least reviewed) step. In contrast, I consider the absence of automated unit tests in the CI workflow more critical.
- Some documentation examples could be made more concise by [limiting or summarizing output to improve readability and reduce scrolling](https://lllangwv.github.io/ParquetDB/01_tutorials/03%20-%20Advanced%20Reads.html#Advanced-Read-Operations).
- The default configuration file is loaded from `config = LoggingConfig.from_yaml(os.path.join(UTILS_DIR, "config.yml"))`. It might be worth considering allowing user-specific configuration files in standard locations (e.g., `~/.config/parquetdb/config.yml` on Linux), for which the [`platformdirs` package](https://pypi.org/project/platformdirs/) could be helpful
- The README currently contains the disclaimer:
  > "Note: This project is in its initial stages. Features and APIs are subject to change. Please refer to the latest documentation and release notes for updates."

  If a JOSS publication is targeted, I would suggest that at least the core API should be considered stable.
- A useful feature might also be to filter by date of creation (if not already possible)

## Own Benchmark

I implemented small benchmark scripts for *ParquetDB*, PyArrow, and SQLite, performing comparable read/write/query operations.

### PyArrow Create

```python
import pyarrow as pa
import pyarrow.parquet as pq
import pyarrow.dataset as ds

dataset = ds.dataset("./local-benchmarks/", format="parquet")

start = 0
for batch in dataset.to_batches(columns=["id"]):
    start = max(batch.column("id").to_numpy().max(), start)

data: List[Dict] = ...

table = pa.Table.from_pylist(data).add_column(0, 'id', [range(start, start + len(data))])
pq.write_table(table, temp_file_path)
```

### ParquetDB Create

```python
from ParquetDB import ParquetDB

db = ParquetDB("./local-benchmarks/")

data: List[Dict] = ...

db.create(data)
```

### SQLite Create

```python

conn = sqlite3.connect("benchmark.sqlite")
cursor = conn.cursor()
cursor.execute("PRAGMA synchronous = OFF")
cursor.execute("PRAGMA journal_mode = MEMORY")
...
cursor.execute("CREATE TABLE ...") # based on data keys
cursor.executemany(sql, values)    # based on data values
conn.commit()
conn.close()
```

### Filter / Query

```python
# _ParquetDB_
db.read(filters=filters)
# PyArrow
dataset.to_table(filter=filter)
# Sqlite
cursor.execute("SELECT * FROM benchmark" + (f" WHERE {query}" if query else ""))
result = cursor.fetchall()
```

with the filter / query being either `None` or a range filter for two columns

```python
filters = [pa.compute.field("column_1") > 100, pa.compute.field("column_97") < 1000]
filter = (pa.compute.field("column_1") > 100) & (pa.compute.field("column_97") < 1000)
query = "column_1 > 100 and column_97 < 1000"
```

### Results

The test data was generated with
```bash
from parquetdb.utils import general_utils

general_utils.generate_pylist_data(n_rows=10**n)
```

Using `%timeit -n1 -r1` (and cleaning of cretead files after each test), I find that

| n = 10ⁿ | Write |  |  | Read |  |  | Query |  |  |
|--------|--------|--------|--------|--------|--------|--------|--------|--------|--------|
|        | _ParquetDB_ | PyArrow | Sqlite | _ParquetDB_ | PyArrow | Sqlite | _ParquetDB_ | PyArrow | Sqlite |
| 2      | 46.1 | 6.25 | 0.949 | 3.96 | 2.6 | 1.63 | 2.95 | 0.885 | 0.29 |
| 3      | 103 | 25 | 6.79 | 3.88 | 2.58 | 13.6 | 4.35 | 2.43 | 1.36 |
| 4      | 709 | 305 | 77.6 | 8.54 | 5.27 | 137 | 9.77 | 6.87 | 5.46 |
| 5      | 7390 | 4890 | 781 | 26.4 | 25.1 | 1360  | 36.4 | 45.2 | 48.2 |

Note: Timings are in milliseconds (ms). Variations for repeated tests were generally in the 1-10% range. The hardware used is slightly less performant than reported in the authors' long paper.

Observations on this system

* SQLite exhibits faster write performance and comparable query performance for small result sets.
* _ParquetDB_ significantly outperforms SQLite when reading large datasets (i.e., returning large arrays).
* A naive PyArrow implementation (replicating basic _ParquetDB_ functionality) often outperforms _ParquetDB_.
