### 2025-04-14 e9a4850
chore: Update .gitignore and refactor input_update_benchmark.py
Added 'github_repos/' to .gitignore to exclude GitHub repository files.
Refactored input_update_benchmark.py for improved readability and
consistency, including formatting changes and the use of
time.perf_counter() for more accurate timing measurements.

No functional changes were made to the benchmarking logic.
