### 2025-04-14 lllangWV/ParquetDB#8 @lllangWV
Hey Christopher,

thanks for opening this issue! I am looking into this now.

Also, I appreciate your review! You gave a lot of useful feedback and suggestions. It will take me a while to write an appropriate response and address some of the points you raised.

Regarding the types tests specifically, these were implemented to validate a the following functionality in ParquetDB: the ability to automatically store and retrieve Python objects (including object instances and Python functions) through serialization with the `dill` library.

It seems the issue arises due to a version conflict in `pyarrow`. This issue is non-trivial to solve, I believe I'll have to open an issue in the arrow repository to track why this is happening.

In the mean time, a temporary fix is forcing users to use `pyarrow==17.0.0`

- `pyarrow==17.0.0`:
Tests runs okay

```bash
Ran 6 tests in 0.035s

OK
```

- `pyarrow==18.1.0`:

The __array__ depreciation warning appears

```bash
...C:\Users\lllang\miniconda3\envs\parquet_review\lib\site-packages\pandas\core\arrays\base.py:725: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and 'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
  return np.array(self, dtype=dtype, copy=copy)
C:\Users\lllang\miniconda3\envs\parquet_review\lib\site-packages\pandas\core\arrays\base.py:725: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and
'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
  return np.array(self, dtype=dtype, copy=copy)
C:\Users\lllang\miniconda3\envs\parquet_review\lib\site-packages\pandas\core\arrays\base.py:725: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and
'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
  return np.array(self, dtype=dtype, copy=copy)
```

- `pyarrow==19.0.1`:
The __array__ depreciation warning still appears and the tests fail

```bash
FAIL: test_arrow_conversion (__main__.TestPythonObjectTypes)
Traceback (most recent call last):
  File "C:\Users\lllang\Desktop\Current_Projects\ParquetDB\tests\test_types.py", line 108, in test_arrow_conversion
    self.assertEqual(arrow_arr.type, PythonObjectArrowType())
AssertionError: DataType(binary) != PythonObjectArrowType(DataType(binary))

Ran 6 tests in 0.045s

FAILED (failures=1)
```

Best,

Logan Lang
