### 2025-04-14 d256190
chore: Pin pyarrow version and remove requirements.txt due to issue raised in issue #8
Updated pyproject.toml to specify pyarrow version 17.0.0 for compatibility.
Removed requirements.txt as dependencies are now managed in pyproject.toml.
