+++ pyproject.toml
@@ -31,7 +31,7 @@ dynamic = ["version"]
 
 dependencies = [
     "python-dotenv",
-    "pyarrow",
+    "pyarrow==17.0.0",
     "pandas",
     "variconfig",
     "matplotlib",
