### 2025-04-17 1b70c95
feat(parquetdb): Add preprocessing method for data without Python objects
Introduced a new static method `preprocess_data_without_python_objects`
to handle incoming data in various formats without using Python objects.
This method improves performance and flexibility when processing data
as it allows for direct conversion from dictionaries, lists, or pandas
dataframes to pyarrow tables.

Updated the `ParquetDB` class to utilize this new method based on the
`serialize_python_objects` flag, enhancing the data processing logic.

No changes were made to the existing functionality when Python objec
[…]
