### 2025-04-17 29581a5
feat(benchmarks): Refactor benchmark scripts to use main function and updated api to use db_path
Updated multiple benchmark scripts to include a main function for betterorganization and readability. This change allows for easier execution andmodularization of the benchmarking logic. The main function initializes the plotting utilities and executes the relevant benchmark functions.

No functional changes were made to the benchmarking logic itself.
