+++ benchmarks/plotting_needle_in_haystack.py
@@ -72,6 +72,20 @@ marker_fill_styles = {
 }
 
 
+def main():
+    from parquetdb.utils import matplotlib_utils
+
+    matplotlib_utils.set_palette("Zissou1")
+
+    # color_diff_log_inset_plot()
+    color_diff_log_inset_plot(
+        savefig=os.path.join(benchmark_dir, "benchmark_needle_in_haystack_times.pdf")
+    )
+    color_diff_log_inset_plot(
+        savefig=os.path.join(benchmark_dir, "benchmark_needle_in_haystack_times.png")
+    )
+
+
 def color_diff_log_inset_plot(savefig=None):
 
     # Initialize a figure and axes
@@ -242,14 +256,4 @@ def color_diff_log_inset_plot(savefig=None):
 
 
 if __name__ == "__main__":
-    from parquetdb.utils import matplotlib_utils
-
-    matplotlib_utils.set_palette("Zissou1")
-
-    # color_diff_log_inset_plot()
-    color_diff_log_inset_plot(
-        savefig=os.path.join(benchmark_dir, "benchmark_needle_in_haystack_times.pdf")
-    )
-    color_diff_log_inset_plot(
-        savefig=os.path.join(benchmark_dir, "benchmark_needle_in_haystack_times.png")
-    )
+    main()
