+++ benchmarks/plotting_update_full_results.py
@@ -59,6 +59,20 @@ marker_fill_styles = {
 }
 
 
+def main():
+    from parquetdb.utils import matplotlib_utils
+
+    matplotlib_utils.set_palette("Zissou1")
+
+    # color_diff_log_inset_plot()
+    color_diff_log_inset_plot(
+        savefig=os.path.join(benchmark_dir, "benchmark_update_full_times.pdf")
+    )
+    color_diff_log_inset_plot(
+        savefig=os.path.join(benchmark_dir, "benchmark_update_full_times.png")
+    )
+
+
 def color_diff_log_inset_plot(savefig=None):
     # Initialize a figure and axes
     fig, ax1 = plt.subplots(figsize=(10, 6))
@@ -214,14 +228,4 @@ def color_diff_log_inset_plot(savefig=None):
 
 
 if __name__ == "__main__":
-    from parquetdb.utils import matplotlib_utils
-
-    matplotlib_utils.set_palette("Zissou1")
-
-    # color_diff_log_inset_plot()
-    color_diff_log_inset_plot(
-        savefig=os.path.join(benchmark_dir, "benchmark_update_full_times.pdf")
-    )
-    color_diff_log_inset_plot(
-        savefig=os.path.join(benchmark_dir, "benchmark_update_full_times.png")
-    )
+    main()
