### 2025-04-19 lllangWV/ParquetDB#11 @perdelt
Docs not up-to-date
I find the documentation well-structured and easy to read. The example output helps me gauge whether everything is working as expected.

I'm going through the examples and noticed that the documentation isn't entirely up-to-date or complete.

- The package has version v0.25.1, and the documentation belongs to v0.24.1.
- The repository's README starts with `db = ParquetDB(dataset_name='parquetdb')`, but apparently `db = ParquetDB('parquetdb')` is correct.
- `NormalizeConfig` is used there, but the import `from parquetdb.core.parquetdb import ParquetDB, NormalizeConfig` is only found in the documentation.
- Same here: https://lllangwv.github.io/ParquetDB/01_tutorials/01%20-%20Getting%20Started.html#Transform-Data
(the normalization example does not work due to missing import)

I suggest checking the examples again. It's convenient for the user if all examples can be run directly.

This issue is part of the JOSS review: openjournals/joss-reviews#7932
