### 2025-04-19 f7a69d2
feat(logging): Implement logging setup and verbosity control
Added a logging configuration to the ParquetDB package to improve debugging and monitoring capabilities. Introduced a `setup_logging` function to initialize logging with different verbosity levels.

Updated the `NormalizeConfig` and `ParquetDB` classes to support logging features, including a new `verbose` parameter for controlling log output. This change enhances the observability of the system during operations.

No changes were made to existing functionality.
