### 2025-04-19 8b15419
refactor(utils): Reorganize imports and improve function formatting
Rearranged import statements for consistency and clarity.
Updated function definitions to follow PEP 8 style guidelines,
including spacing around parameters and improved readability of function bodies. This change enhances code maintainability and adherence to Python standards.

No functional changes were made to the existing logic.
