### 2025-04-19 e64ba5d
refactor(config): Replace LoggingConfig with ConfigDict and update parameters
Replaced the LoggingConfig import with ConfigDict to streamline configuration handling.
Updated the `max_rows_per_file` and `min_rows_per_group` parameters in the configuration YAML file to accommodate larger datasets and improve performance.

No functional changes were made to the existing logic.
