+++ parquetdb/utils/config.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 
 from dotenv import load_dotenv
-from variconfig import LoggingConfig
+from variconfig import ConfigDict
 
 load_dotenv()
 
@@ -12,7 +12,7 @@ UTILS_DIR = str(FILE.parents[0])
 DATA_DIR = os.getenv("DATA_DIR")
 
 
-config = LoggingConfig.from_yaml(os.path.join(UTILS_DIR, "config.yml"))
+config = ConfigDict.from_yaml(os.path.join(UTILS_DIR, "config.yml"))
 
 if DATA_DIR:
     config.data_dir = DATA_DIR
