### 2025-04-19 ea678b1
chore: Remove obsolete benchmark scripts for plotting
Deleted several benchmark scripts related to plotting input updates, needle-in-haystack, read/write, and update benchmarks. These scripts are no longer needed as the functionality has been migrated to Jupyter notebooks for better organization and usability.

No functional changes were made to the remaining codebase.
