### 2025-04-19 8bceffd
feat(docs): Add benchmark gallery and update example gallery links
Introduced a new Benchmark Gallery section in the documentation to showcase various performance benchmarks. Updated the Example Gallery to include links to the newly added benchmarks, enhancing the documentation's comprehensiveness and usability.

No functional changes were made to the codebase.
