### 2025-04-19 lllangWV/ParquetDB#11 @lllangWV
Hey,

Thanks for taking the time to review the package and give feedback!

I have addressed the first three points you raised. I am looking into a github actions to automatically build and update the version number in the docs on a new release of the package.

For the fourth point, what exactly were you referencing for it? I clicked the link, and it looks fine to me.

I changed how import can be done, now they can just directly import them from the package.

```python
from parquetdb import ParquetDB, NormalizeConfig
```

> I suggest checking the examples again. It's convenient for the user if all examples can be run directly.

Do you suggest, I add `pymongo`, `pymatgen`, and `ipython` to the core dependencies?  Originally, I did not include them because they are not directly related to the package. I can add all the dependencies needed such that they can immediately run the example, if you think that will be better.

Logan Lang
