### 2025-04-19 a0a6e27
feat(docs): Update documentation links and add benchmark examples
Updated various documentation files to reflect changes in the benchmark gallery and example sections. Adjusted links to ensure they point to the newly added benchmark examples, enhancing the usability of the documentation.

Additionally, modified import statements in tutorial notebooks to streamline the code and improve clarity.

No functional changes were made to the codebase.
