+++ docs/_sources/examples/index.rst.txt
@@ -7,7 +7,7 @@ These examples are automatically generated from the `examples
 directory`_ of the package and showcase how to effectively use ParquetDB's features for data storage, querying, and management. Feel free to download and run these notebooks to explore the functionality firsthand.
 
 .. _examples directory: https://github.com/lllangWV/ParquetDB/tree/main/examples/notebooks
-
+.. _benchmarks directory: https://github.com/lllangWV/ParquetDB/tree/main/examples/notebooks/benchmarks
 
 .. nblinkgallery::
     :caption: Example Gallery
@@ -18,14 +18,34 @@ directory`_ of the package and showcase how to effectively use ParquetDB's featu
     notebooks/Example 3 - Jarvis-dft_3d
     notebooks/Example 4 - Jarvis-dft_pbe_2d
 
+.. nblinkgallery::
+    :caption: Benchmark Gallery
+    :name: rst-benchmark-gallery
+
+    benchmarks/Needle-in-Haystack Benchmark
+    benchmarks/Input Update Benchmark
+    benchmarks/Create and Read Benchmark
+    benchmarks/Update Into Constant Rows Benchmark
+    benchmarks/Full DB Updates Benchmark
+
 Contents
 --------
 
 .. toctree::
-   :maxdepth: 3
+   :maxdepth: 2
    :caption: Example Gallery
 
    notebooks/Example 1 - 3D Alexandria Database
    notebooks/Example 2 - 3D Alexandria Advanced
    notebooks/Example 3 - Jarvis-dft_3d
    notebooks/Example 4 - Jarvis-dft_pbe_2d
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Benchmark Gallery
+
+   benchmarks/Needle-in-Haystack Benchmark
+   benchmarks/Input Update Benchmark
+   benchmarks/Create and Read Benchmark
+   benchmarks/Update Into Constant Rows Benchmark
+   benchmarks/Full DB Updates Benchmark
