+++ docs/_sources/examples/benchmarks/Needle-in-Haystack Benchmark.ipynb.txt
@@ -1,0 +1,610 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Needle-in-Haystack Benchmark Notebook\n",
+    "\n",
+    "This notebook measures the performance of a “needle in a haystack” lookup query—finding one specific row among many—for three storage backends: SQLite, MongoDB, and ParquetDB.\n",
+    "\n",
+    "---\n",
+    "\n",
+    "## Benchmark Details\n",
+    "\n",
+    "- **Data Generation:**  \n",
+    "  - Each run generates *N* rows × 100 columns of integers, with exactly one row set to `–1` (“the needle”) and all other entries random in [0, 1 000 000].  \n",
+    "  - Row counts tested: `1, 10, 100, 1 000, 10 000, 100 000, 1 000 000`.  \n",
+    "  - Integers chosen as a basic primitive type—byte size is the main factor, so these results represent a **lower bound** on query time. More complex or larger types will incur higher cost.\n",
+    "\n",
+    "- **Parquet Normalization Settings (defaults):**  \n",
+    "  - **Row‐group size:** 50 000–100 000 rows per group  \n",
+    "  - **Max rows per file:** 10 000 000  \n",
+    "  - Tuning these can shift performance between inserts, reads, and updates.\n",
+    "\n",
+    "---\n",
+    "\n",
+    "## System Specifications\n",
+    "\n",
+    "- **OS:** Windows 10  \n",
+    "- **CPU:** AMD Ryzen 7 3700X @ 3.6 GHz (8 physical cores, 16 logical)  \n",
+    "- **RAM:** 128 GB DDR4‑3600 MHz (4×32 GB)  \n",
+    "\n",
+    "---"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## 1. Setup"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import os\n",
+    "import time\n",
+    "import random\n",
+    "import shutil\n",
+    "import sqlite3\n",
+    "import pandas as pd\n",
+    "import matplotlib.pyplot as plt\n",
+    "import matplotlib.ticker as ticker\n",
+    "from mpl_toolkits.axes_grid1.inset_locator import inset_axes\n",
+    "from pymongo import MongoClient\n",
+    "import pyarrow.compute as pc\n",
+    "\n",
+    "from parquetdb import ParquetDB, config\n",
+    "\n",
+    "\n",
+    "bench_dir = os.path.join(config.data_dir, \"benchmarks\")\n",
+    "sqlite_dir = os.path.join(bench_dir, \"sqlite\")\n",
+    "mongo_dir  = os.path.join(bench_dir, \"mongodb\")\n",
+    "pq_dir     = os.path.join(bench_dir, \"parquetdb\")\n",
+    "for d in (sqlite_dir, mongo_dir, pq_dir):\n",
+    "    os.makedirs(d, exist_ok=True)\n",
+    "\n",
+    "row_counts = [1, 10, 100, 1_000, 10_000, 100_000, 1_000_000]\n",
+    "n_cols = 100"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## 2. SQLite Becnhmark"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### 2.1 Helper functions"
+   ]
+  },
+  {
[…]
