### 2025-04-20 c1a6297
refactor(tests): Update test files for improved readability and structure
Refactored test files by removing commented-out code and unnecessary logging configurations. Introduced a `VERBOSE` variable for better control over logging verbosity in tests. Updated the instantiation of `ParquetDB` to include the `verbose` parameter.

These changes enhance the clarity and maintainability of the test code without altering the existing test functionality.
