### 2025-04-20 50632ea
feat(parquetdb): Add logging for data preprocessing methods
Introduced logging statements in the preprocess_data_without_python_objects
and preprocess_data_with_python_objects methods to enhance traceability during data processing. This change allows for better monitoring of the data handling process and aids in debugging.

No functional changes were made to the existing logic.
