### 2025-04-21 01c7d25
feat(docs): Add Small Benchmark example to documentation
Included a new Small Benchmark example in the documentation to demonstrate
performance comparisons between ParquetDB, PyArrow, and SQLite. This addition
provides users with practical insights into the read/write/query operations
of the database.

Updated the index.rst file to include a reference to the new benchmark
notebook, enhancing the overall documentation structure.
