### 2025-04-21 1ff6854
feat(utils): Add prefix parameter to generate_pylist_data function
Introduced a new parameter `prefix` to the `generate_pylist_data` function
to allow customization of column names in the generated data. This change
enhances the flexibility of the function, enabling users to specify a
different prefix for the columns instead of the default "column_".

No functional changes were made to the existing logic.
