### 2025-04-21 b8d5bd3
fix(logging): Remove file handler from logging configuration
Updated the logging configuration to remove the file handler from the
ParquetDB logger. This change simplifies the logging setup by relying
only on the console handler, which is sufficient for current logging
needs.

No functional changes were made to the existing logic.
