### 2025-04-21 lllangWV/ParquetDB#9 @lllangWV
Hey [Christopher](https://github.com/ckoerber),

Thanks for pointing of the depreciation of the initialization arguments throughout the repository.

I have updated the benchmarks, the arxiv paper, and documentation.

Logan Lang
