### 2025-04-21 faab446
fix(parquetdb): Correct save path calculation for parquet files
Updated the logic for determining the save path of parquet files to
ensure that the next file index is correctly calculated. The previous
implementation used the number of existing files, which could lead to
overwriting files. Now, the maximum index from existing files is
determined and incremented for the new save path.

This change prevents potential data loss by ensuring unique file names.
