### 2025-04-21 lllangWV/ParquetDB#10 @lllangWV
Hey [Christopher](https://github.com/ckoerber),

I just pushed a commit that address this issue.

I appreciate the suggestions on how to manage it. I have considered option 3 before when first deciding what to do it. However, I decided against it as it is not compatible with the `pyarrow.dataset.write_dataset` method which is used in the normalization of the database.

I decided to iterate through the files and determine the max number and just increment it by 1. While this is obviously slower than just using the number of files, it is not slower by much. This will only become an an issue if you have millions of files, and at that point the user should probably normalize the database to condense it into a smaller number of files.

For instance, I have used ParuqetDB with the 1000Genomes dataset which has ~84 million rows and 5000 columns.  At worst I had 84 files with the with million rows each. However, after normalized the database down to 5 files with 20 million rows each. In the normalized dataset the file sizes ~3GB each with row groups having a size of ~300MB.

In summary for practical purposes, I doubt the using a slower method for getting the maximum index number will be an issue.
