### 2025-04-22 bef4bbe
chore: Remove obsolete environment YAML files and update .gitignore
Deleted unused environment configuration files: env_dev.yml,
env_material.yml, and env.yml. These files were no longer needed
for the project setup. Updated .gitignore to include new environment
files to prevent them from being tracked in the repository.

This change helps maintain a cleaner project structure.
