+++ env.yml
@@ -1,9 +1,0 @@
-name: parquetdb
-channels:
-- defaults
-dependencies:
-- python==3.9.*
-- pip
-- pip:
-  - .
-
