### 2025-04-22 lllangWV/ParquetDB#11 @perdelt
Hey,

the package itself seems to be well-built. So I don't think there's anything that should be included in the requirements that would be helpful for comparisons but isn't necessary for the package. It would be nice if the example pages included instructions on how to install the packages used there, for the convenience of the user.

> For the fourth point, what exactly were you referencing for it? I clicked the link, and it looks fine to me.

Nevermind this point, the example is perfect as is. NormalizeConfig is not necessary to be included here, only for customization, my fault.
