### 2025-04-22 a5bd798
feat(docs): Update Sphinx configuration and add notebook CSS
Modified the Sphinx configuration to improve documentation formatting.
Updated the extensions list for consistency and added a new CSS file
to control the output area height in Jupyter notebooks, enhancing
readability.

These changes aim to provide a better user experience when viewing
documentation and notebook outputs.
