### 2025-04-22 63a3896
feat(docs): Add notebook CSS and installation commands for libraries
Included a new CSS file for Jupyter notebooks to control output area
height, enhancing readability. Additionally, added installation commands
for `parquetdb` and `pymongo` in various benchmark and notebook examples
to ensure users have the necessary dependencies.

These changes aim to improve user experience and facilitate smoother
execution of the notebooks and benchmarks.
