### 2025-04-22 b96fa52
refactor: Reorganize imports, improve code formatting changed model to o4-mini
Adjusted the order of imports for better readability and consistency.
Updated variable assignments and function calls to follow a more
uniform style. These changes aim to enhance code maintainability
and clarity without altering functionality.
