### 2025-04-24 321c1bc
feat(workflow): Add scripts for summarizing PR commits and comments
Introduced two new Python scripts to automate the summarization of
commit messages and PR comments for releases. The `summarize_pr_commits.py`
script extracts commit messages from a merged PR and classifies them into
categories, while `summarize_release-pr_comments.py` summarizes comments
from multiple PRs into a comprehensive release summary.

These additions aim to streamline the release process by providing
concise summaries, improving clarity for release notes, and enhancing
the overall workflow efficiency.
