+++ .github/workflows/test-pull-requests.yml
@@ -1,0 +1,48 @@
+name: Run Tests
+
+on:
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  test:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false  # This allows all test environments to complete even if one fails
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ['3.8', '3.13']
+
+    steps:
+    - uses: actions/checkout@v3
+      with:
+        fetch-depth: 0  # Fetches all history for proper pytest coverage
+    
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+        
+    - name: Cache pip dependencies
+      uses: actions/cache@v3
+      with:
+        path: ~/.cache/pip
+        key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}-${{ hashFiles('**/setup.py') }}
+        restore-keys: |
+          ${{ runner.os }}-pip-
+          
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install pytest pytest-cov
+        # Install the package itself if it has a setup.py
+        if [ -f setup.py ]; then pip install -e .; fi
+        # Install requirements if present
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        # Install test requirements if present
+        if [ -f test-requirements.txt ]; then pip install -r test-requirements.txt; fi
+      shell: bash
+        
+    - name: Test with pytest
+      run: |
+        pytest tests/ -v
