### 2025-04-24 254d586
feat(docs): Add Read the Docs configuration file
Introduced a new configuration file for Read the Docs to define the
build environment and Sphinx settings for documentation. This change
aims to facilitate automated documentation builds and ensure
consistency in the documentation process.

Also updated the `pyproject.toml` file to remove an unnecessary line
break in the documentation dependencies list.
