+++ pyproject.toml
@@ -66,7 +66,7 @@ docs= [
     "sphinxcontrib-video",
     "pandoc",
     "furo",
-    "numpydoc",
+    "numpydoc"
 ]
 
 
