### 2025-04-24 fae31e2
feat(tests): Add test dependencies to pyproject.toml and update workflow
Included `pytest` and `pytest-cov` in the test dependencies section of
`pyproject.toml` to ensure proper testing setup. Updated the GitHub
Actions workflow to install these dependencies correctly by modifying
the installation command to include the `[docs]` extras.

These changes aim to streamline the testing process and ensure all
necessary tools are available for running tests.
