### 2025-04-24 d33317f
fix(workflow): Update test dependencies installation command
Modified the GitHub Actions workflow to install test dependencies
using the `[tests]` extras instead of `[docs]`. This change ensures
that the correct dependencies are installed for running tests,
streamlining the testing process and improving reliability.
