+++ .github/workflows/test-pull-requests.yml
@@ -34,13 +34,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pytest pytest-cov
-        # Install the package itself if it has a setup.py
-        if [ -f setup.py ]; then pip install -e .[docs]; fi
-        # Install requirements if present
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        # Install test requirements if present
-        if [ -f test-requirements.txt ]; then pip install -r test-requirements.txt; fi
+        if [ -f setup.py ]; then pip install -e .[tests]; fi
+  
       shell: bash
         
     - name: Test with pytest
