### 2025-04-24 1d643ce
fix(workflow): Update Python version in test workflow
Changed the Python version in the GitHub Actions workflow from 3.13
to 3.12 to align with the supported versions for testing. This update
ensures compatibility with the current project requirements and
maintains the integrity of the testing process.
