### 2025-04-25 342133d
fix(parquetdb): Replace print statements with logger for better tracking
Replaced print statements with logger calls to improve logging
consistency and facilitate debugging. This change allows for
better tracking of the transformation process and file handling
within the ParquetDB class, enhancing overall maintainability.
