### 2025-04-25 b408b9d
fix(parquetdb): Add debug logging for schema and table updates
Introduced debug logging to track incoming, current, and merged column
names during schema updates. This change enhances visibility into the
data transformation process, aiding in debugging and ensuring data
integrity. Additionally, sorted the update table and current table
by ID to maintain order during updates.
