### 2025-04-25 70366de
fix(parquetdb): Remove redundant debug logging for update table shape
Eliminated the debug log statement that printed the shape of the
update table, as it was deemed unnecessary. This change aims to
reduce log clutter and improve the clarity of debug information
during table updates.
