### 2025-04-25 15b3a75
fix(tests): Improve temporary directory management in tests
Updated the test suite to use a single temporary directory for
database tests, reducing the risk of race conditions. Enhanced
the tearDown method to check for the existence of temporary
directories before attempting to remove them, ensuring proper
cleanup. Added debug logging to track directory contents during
test execution, improving visibility into the test environment.
