+++ tests/test_parquetdb.py
@@ -23,12 +23,14 @@ VERBOSE = 2
 with open(os.path.join(config.tests_dir, "data", "alexandria_test.json"), "r") as f:
     alexandria_data = json.load(f)
 
+TEMP_DIR = tempfile.mkdtemp()
+
 
 class TestParquetDB(unittest.TestCase):
     def setUp(self):
         # Create a temporary directory for the database
-        self.temp_dir = tempfile.mkdtemp()
-        self.temp_dir_2 = tempfile.mkdtemp()
+        self.temp_dir = os.path.join(TEMP_DIR, "temp-1")
+        self.temp_dir_2 = os.path.join(TEMP_DIR, "temp-2")
 
         self.db = ParquetDB(
             db_path=os.path.join(self.temp_dir, "test_db"), verbose=VERBOSE
@@ -46,7 +48,11 @@ class TestParquetDB(unittest.TestCase):
     def tearDown(self):
         # Remove the temporary directory after the test
         try:
-            shutil.rmtree(self.temp_dir)
+            if os.path.exists(self.temp_dir):
+                shutil.rmtree(self.temp_dir)
+
+            if os.path.exists(self.temp_dir_2):
+                shutil.rmtree(self.temp_dir_2)
         except Exception as e:
             logger.error(f"Error during cleanup: {e}")
         # For some reason, there are race conditions when
@@ -656,10 +662,17 @@ class TestParquetDB(unittest.TestCase):
         data_2 = [{"id": 0, "density": 5}]
 
         self.db.update(data_2)
-        table = self.db.read()
+        table = self.db.read(ids=[1])
+        df = table.to_pandas()
+        logger.debug(f"Dataframe: \n {df}")
+        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [
+            [0, 1, 0],
+        ]
+
+        table = self.db.read(ids=[0])
+
         assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [
             [1, 0, 0],
-            [0, 1, 0],
         ]
 
     def test_update_maintains_existing_extension_arrays_batches(self):
@@ -705,6 +718,8 @@ class TestParquetDB(unittest.TestCase):
 
         self.db.update(data_2, update_keys="material_id")
         table = self.db.read()
+        df = table.to_pandas()
+        logger.debug(f"Dataframe: \n {df}")
         assert table["material_id"].combine_chunks().to_pylist() == [1, 2]
         assert table["material_name"].combine_chunks().to_pylist() == [
             "material_1_updated",
@@ -775,6 +790,7 @@ class TestParquetDB(unittest.TestCase):
         table = self.db.read()
         df = table.to_pandas()
         logger.debug(f"Dataframe: \n {df}")
+
         assert table["field_1"].combine_chunks().to_pylist() == [
             "here",
             None,
@@ -822,6 +838,7 @@ class TestParquetDB(unittest.TestCase):
 
         incoming_table = ParquetDB.construct_table(incoming_data)
 
+        logger.info(f"Starting test on a left outer join")
         join_type = "left outer"
         left_outer_table_pyarrow = incoming_table.join(
             current_table,
@@ -849,13 +866,20 @@ class TestParquetDB(unittest.TestCase):
         column_names = left_outer_table.column_names
         names_sorted = sorted(column_names)
         left_outer_table_sorted = left_outer_table.select(names_sorted)
-
+        left_outer_table_pyarrow_sorted = left_outer_table_pyarrow_sorted.sort_by(
+            [("id_1", "ascending"), ("id_2", "ascending")]
+        )
+        left_outer_table_sorted = left_outer_table_sorted.sort_by(
+            [("id_1", "ascending"), ("id_2", "ascending")]
+        )
         for name in left_outer_table_sorted.column_names:
-            assert (
[…]
