### 2025-04-25 0c67d57
fix(docs): Correct indentation in Read the Docs configuration
Adjusted the indentation in the .readthedocs.yaml file to ensure
proper parsing of the Python installation requirements. This change
aims to maintain the integrity of the documentation build process
by aligning with YAML formatting standards.
