+++ .readthedocs.yaml
@@ -18,8 +18,8 @@ sphinx:
 # declare the Python requirements required to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
-   install:
-   - method: pip
+  install:
+    - method: pip
       path: .
       extra_requirements:
         - docs
