### 2025-04-25 9eb6691
fix(docs): Update documentation files
Moving to readthedocs instead of github pages. This will allow for versioning documentation. This means we do not have to store the build docs inside the repo as readthedocs handles this. Now moving the source to docs directory
