+++ docs/_sources/01_tutorials/04 - Supported Data Types.ipynb.txt
@@ -1,603 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# Supported Data Types\n",
-    "\n",
-    "ParquetDB is designed to handle a variety of data types, making it flexible for diverse applications. This includes:\n",
-    "- **Empty dictionaries**\n",
-    "- **Nested dictionaries**\n",
-    "- **Lists**\n",
-    "- **NumPy arrays**\n",
-    "- **Python functions** \n",
-    "- **Class instances (objects)**\n",
-    "\n",
-    "In this notebook, we’ll demonstrate how to create (or update) records in ParquetDB with these data types. \n"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 27,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import pprint\n",
-    "import shutil\n",
-    "import os\n",
-    "import pandas as pd\n",
-    "import pyarrow as pa\n",
-    "from parquetdb import ParquetDB, NormalizeConfig\n",
-    "\n",
-    "db_path = \"ParquetDB\"\n",
-    "if os.path.exists(db_path):\n",
-    "    shutil.rmtree(db_path)\n",
-    "\n",
-    "db = ParquetDB(db_path)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "## Empty Dictionaries\n",
-    "\n",
-    "Empty dictionaries can be part of your records without issues. This can be useful for placeholders or optional fields. \n",
-    "\n",
-    "Parquet files cannot natively handle empty dictionaries, so ParquetDB handles them by creating a dummy field with a null value."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 28,
-   "metadata": {},
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "   empty_dict.dummy_field  id\n",
-      "0                     NaN   0\n"
-     ]
-    }
-   ],
-   "source": [
-    "data_with_empty_dict = [\n",
-    "    {\n",
-    "        \"empty_dict\": {},\n",
-    "    }\n",
-    "]\n",
-    "\n",
-    "db.create(data_with_empty_dict)\n",
-    "\n",
-    "df = db.read().to_pandas()\n",
-    "print(df)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "## Nested Dictionaries\n",
-    "\n",
-    "You can store nested dictionaries directly. ParuqetDB will flatten the nested dictionaries over multiple columns, by giving child fields the following syntax:\n",
-    "\n",
-    "```\n",
-    "parent_field_name.child_field_name\n",
-    "```\n",
-    "\n"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 29,
-   "metadata": {},
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "   empty_dict.dummy_field  id nested_dict.level1_key  \\\n",
-      "0                     NaN   0                   None   \n",
-      "1                     NaN   1                 value1   \n",
-      "\n",
-      "   nested_dict.level1_subdict.level2_key  \\\n",
-      "0                                    NaN   \n",
-      "1                                  123.0   \n",
-      "\n",
-      "  nested_dict.level1_subdict.level2_subdict.level3_key  \n",
-      "0                                               None    \n",
-      "1                                               True    \n"
-     ]
-    }
-   ],
-   "source": [
-    "data_with_nested_dict = [\n",
-    "    {\n",
[…]

[…]
