+++ sphinx/source/conf.py
@@ -1,193 +1,0 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
-
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-import os
-import shutil
-import sys
-from distutils.sysconfig import get_python_lib
-
-from parquetdb._version import version
-
-# ---------------
-project = "ParquetDB"
-copyright = "2024, Logan Lang"
-author = "Logan Lang"
-
-
-sys.path.insert(0, os.path.abspath("."))
-
-repo_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))
-src_examples_path = os.path.abspath(os.path.join(os.path.dirname(__file__), "examples"))
-if os.path.exists(src_examples_path):
-    shutil.rmtree(src_examples_path)
-
-shutil.copytree(os.path.join(repo_root, "examples"), src_examples_path)
-print(repo_root)
-print(src_examples_path)
-
-# examples_path = os.path.join(repo_root, "examples")
-# sys.path.insert(0, examples_path)
-
-
-if os.environ.get("READTHEDOCS") == "True":
-
-    site_path = get_python_lib()
-    ffmpeg_path = os.path.join(site_path, "imageio_ffmpeg", "binaries")
-    print("########")
-    print("good1")
-    [ffmpeg_bin] = [
-        file for file in os.listdir(ffmpeg_path) if file.startswith("ffmpeg-")
-    ]
-    print("########*****")
-    print("good2")
-    try:
-        os.symlink(
-            os.path.join(ffmpeg_path, ffmpeg_bin), os.path.join(ffmpeg_path, "ffmpeg")
-        )
-    except FileExistsError:
-        print("File is already there!!!!!!!")
-    else:
-        print("file created :)")
-    print("good3")
-    os.environ["PATH"] += os.pathsep + ffmpeg_path
-    print("good4")
-
-# sys.path.append(os.path.abspath("./_ext"))
-# -- Project information -----------------------------------------------------
-
-
-# The full version, including alpha/beta/rc tags
-release = version
-
-# -- General configuration ---------------------------------------------------
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
-extensions = [
-    "sphinx_copybutton",
-    "sphinx.ext.autodoc",
-    "sphinx.ext.autosummary",
-    "nbsphinx",
-    "sphinx_new_tab_link",
-    "sphinx.ext.intersphinx",
-    "sphinxcontrib.youtube",
-    "sphinxcontrib.video",
-    "sphinx_design",
-    "sphinx.ext.githubpages",
-    "sphinx.ext.napoleon",
-    "numpydoc",
-    "sphinx.ext.autodoc",
-    "sphinx.ext.autosummary",
-    # "sphinx-nbexamples",
-    # "sphinx_gallery.gen_gallery",
-    # 'sphinx.youtube',
-    # 'monokai_colors.ManimMonokaiStyle'
-]
-nbsphinx_allow_errors = True
-pygments_style = "sphinx"
-
-
-# sphinx_gallery_conf = {
-#     # convert rst to md for ipynb
-#     "pypandoc": True,
-#     # path to your examples scripts
-#     "examples_dirs": ["../../examples/notebooks"],
-#     # path where to save gallery generated examples
-#     "gallery_dirs": ["examples"],
-#     "doc_module": "parquetdb",
[…]
