+++ sphinx/source/03_api/core/_autosummary/parquetdb.core.parquetdb.NormalizeConfig.rst
@@ -1,45 +1,0 @@
-﻿parquetdb.core.parquetdb.NormalizeConfig
-========================================
-
-.. currentmodule:: parquetdb.core.parquetdb
-
-.. autoclass:: NormalizeConfig
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~NormalizeConfig.__init__
-   
-   
-
-   
-   
-   .. rubric:: Attributes
-
-   .. autosummary::
-   
-      ~NormalizeConfig.batch_readahead
-      ~NormalizeConfig.batch_size
-      ~NormalizeConfig.create_dir
-      ~NormalizeConfig.existing_data_behavior
-      ~NormalizeConfig.file_options
-      ~NormalizeConfig.file_visitor
-      ~NormalizeConfig.fileformat
-      ~NormalizeConfig.filesystem
-      ~NormalizeConfig.fragment_readahead
-      ~NormalizeConfig.fragment_scan_options
-      ~NormalizeConfig.load_format
-      ~NormalizeConfig.max_open_files
-      ~NormalizeConfig.max_partitions
-      ~NormalizeConfig.max_rows_per_file
-      ~NormalizeConfig.max_rows_per_group
-      ~NormalizeConfig.memory_pool
-      ~NormalizeConfig.min_rows_per_group
-      ~NormalizeConfig.use_threads
-   
-   
