+++ sphinx/source/03_api/core/normalize_config.rst
@@ -1,9 +1,0 @@
-NormalizeConfig
-========================
-
-- :class:`NormalizeConfig <parquetdb.core.parquetdb.NormalizeConfig>` - A configuration dataclass that controls how data normalization operations are performed in the database. It provides settings for optimizing performance through row distribution, file structure, batch processing, and memory management. Parameters include batch sizes, read-ahead settings, threading options, and file organization constraints.
-
-.. autosummary::
-   :toctree: _autosummary
-
-   parquetdb.core.parquetdb.NormalizeConfig
