+++ sphinx/source/03_api/utils/pyarrow_utils.rst
@@ -1,27 +1,0 @@
-PyArrowUtils
-========================
-
-
-.. currentmodule:: parquetdb.utils.pyarrow_utils
-
-.. autosummary::
-   :nosignatures:
-
-   is_extension_type
-   is_fixed_shape_tensor
-   drop_duplicates
-   join_tables
-   sort_schema
-   table_schema_cast
-   unify_schemas
-   delete_columns
-   delete_ids
-   delete_field_values
-
-Detailed Documentation
------------------------
-
-.. automodule:: parquetdb.utils.pyarrow_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
