### 2025-04-25 75bf9da
fix(docs): Update documentation link to point to index.html
Changed the documentation link in README.md to direct users to the
index.html file instead of the root URL. This ensures that users
access the correct starting point of the documentation, improving
navigation and usability.
