### 2025-04-25 eb44ade
chore: Remove obsolete version and initialization files
Deleted __init__.py and _version.py as they are no longer needed.
This cleanup helps streamline the codebase and eliminates unused
files that may cause confusion.
