+++ __init__.py
@@ -1,5 +1,0 @@
-from parquetdb._version import __version__
-
-from parquetdb.utils.config import config
-
-from parquetdb.core.parquetdb import ParquetDB
