### 2025-04-25 c68b5fa
chore: Remove deprecated workflow and utility scripts
Deleted several obsolete scripts related to GitHub workflows and utilities
that are no longer in use. This cleanup helps streamline the codebase and
eliminates confusion from unused files.
