### 2025-04-25 906f981
docs: Revise contributing guidelines and update development dependencies
Updated the CONTRIBUTING.md file to enhance clarity on contribution
processes, including setting up a development environment and running
tests. Removed the outdated CONTRIBUTING.rst file.

Also, modified the pyproject.toml to include additional development
dependencies for improved testing and documentation generation.
