### 2025-04-25 fccd541
chore: Rename test job and add documentation build test workflow
Renamed the test job to 'run-tests' for clarity and added a new
workflow to build documentation using Sphinx. The new workflow
includes steps for setting up Python, installing dependencies,
and checking the build status, ensuring that documentation
is properly generated and validated during CI.
