+++ .github/workflows/test-pull-requests.yml
@@ -5,7 +5,7 @@ on:
     branches: [ main ]
 
 jobs:
-  test:
+  run-tests:
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false  # This allows all test environments to complete even if one fails
@@ -41,3 +41,40 @@ jobs:
     - name: Test with pytest
       run: |
         pytest tests/ -v
+
+  run-test-build-docs:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0  # Fetch all history
+
+      - name: Set up Python
+        uses: actions/setup-python@v3
+        with:
+          python-version: '3.x'
+
+      - name: Install pandoc
+        uses: pandoc/actions/setup@v1
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install requests build python-dotenv
+          pip install setuptools  # Add setuptools which provides distutils
+          pip install .[docs]  # Install package with docs dependencies
+
+      - name: Build documentation
+        id: build_docs
+        continue-on-error: true
+        run: |
+          cd docs  # Assuming Sphinx docs are in this directory
+          make clean && make html
+          echo "build_successful=true" >> $GITHUB_OUTPUT
+
+      - name: Check build status
+        if: steps.build_docs.outputs.build_successful != 'true'
+        run: |
+          echo "Documentation build failed!"
+          exit 1
+          - name: Delete release if failed
