### 2025-04-25 03fd3ac
chore: Update Python version in test workflow to 3.10
Changed the Python version in the test workflow from '3.x' to '3.10'
to ensure compatibility with the latest features and improvements.
This update helps maintain a consistent development environment
across CI processes.
