+++ .github/workflows/test-pull-requests.yml
@@ -52,7 +52,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v3
         with:
-          python-version: '3.x'
+          python-version: '3.10'
 
       - name: Install pandoc
         uses: pandoc/actions/setup@v1
