### 2025-04-25 4877633
chore: Change continue-on-error setting for documentation build
Updated the continue-on-error setting in the documentation build step
to false. This change ensures that the workflow fails if the documentation
build encounters errors, allowing for better error tracking and resolution
during CI processes.
