+++ .github/workflows/test-pull-requests.yml
@@ -66,7 +66,7 @@ jobs:
 
       - name: Build documentation
         id: build_docs
-        continue-on-error: true
+        continue-on-error: false
         run: |
           cd docs  # Assuming Sphinx docs are in this directory
           make clean && make html
