### 2025-04-25 38f4db5
chore: Remove build status check from documentation workflow
Eliminated the build status check step in the documentation build
workflow. This change simplifies the workflow by removing unnecessary
failure conditions, allowing for a more streamlined documentation
generation process.
