### 2025-04-25 8dd22bc
chore: Add test-build-package job to CI workflow
Introduced a new job in the CI workflow to build and publish the package
upon release. This job includes steps for checking out the repository,
setting up Python, installing dependencies, and executing the build.
It ensures that the package is built and published automatically,
streamlining the release process.
