### 2025-04-25 fe8f7b1
chore: Simplify test workflow by removing unnecessary steps
Removed echo commands and the publish package step from the test
workflow to streamline the process. Updated the Python version to
3.10 for consistency and ensured the build command runs directly
without additional environment variables.

This change helps maintain a cleaner and more efficient CI workflow.
