+++ .github/workflows/test-pull-requests.yml
@@ -85,27 +85,14 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
-        python-version: '3.x'
+        python-version: '3.10'
 
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
 
-    - name: Echo release id
-      run: |
-      
-        echo "RELEASE_ID=${{ github.event.release.id }}"
-        echo "REPO_NAME=${{ github.repository}}"
-        echo "TAG=${{ github.event.release.tag_name }}"
-    
+
     - name: Test Build 
-      env:
-        SETUPTOOLS_SCM_PRETEND_VERSION: ${{ github.event.release.tag_name }} 
-      run: python -m build
- 
-    - name: Publish package
-      if: success()
-      uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
-      with:
-        user: __token__
-        password: ${{ secrets.PYPI_API_TOKEN }}
+      run: |
+        python -m build
+
