### 2025-04-25 230aecc
chore: Add GitHub Actions workflow for creating release branches
Introduced a new workflow that triggers on published releases to
automatically create or update a branch corresponding to the release
tag. This workflow includes steps for checking out the repository,
setting up Python, installing dependencies, building the package,
and publishing it to PyPI.

This change streamlines the release process by ensuring that
branches are consistently created for each release.
