### 2025-04-25 898b3b5
chore: Install build package in CI workflows
Added the installation of the 'build' package in both the release and
test workflows to ensure that the package can be built correctly. This
change supports the automation of the build process during CI,
streamlining the release and testing procedures.
