+++ .github/workflows/on-release.yml
@@ -24,6 +24,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install build
 
     - name: Echo release id
       run: |
