+++ .github/workflows/test-pull-requests.yml
@@ -90,6 +90,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install build
 
 
     - name: Test Build 
