### 2025-04-25 8114902
chore: Update dependencies and refactor configuration loading
Removed trailing space from the 'distributed' dependency in pyproject.toml
and added 'platformdirs' for better configuration management.

Refactored the configuration loading in config.py to create a user-specific
config file if it doesn't exist, enhancing the flexibility of the package
configuration.

This change improves the overall setup process for users and ensures
that configuration files are properly managed.
