### 2025-04-28 903b3e8
chore: Add OpenAI, python-dotenv, and PyGithub to development dependencies
Included 'openai', 'python-dotenv', and 'PyGithub' in the development
dependencies of pyproject.toml to support new features and enhance
development workflows. This change ensures that necessary libraries
are available for development and testing.
