### 2025-04-28 304da33
chore: Remove initial project note from README
Deleted the note indicating that the project is in its initial stages.
This change reflects the project's maturity and reduces clutter in the
documentation.
