### 2025-04-28 e47de7e
Update CHANGELOG for v0.26.0
