### 2025-04-28 7e1af2e
chore: Bump version to 0.26.0
Updated the version number in _version.py to reflect the new release.
This change is part of the preparation for the upcoming v0.26.0 release.
