+++ parquetdb/_version.py
@@ -17,5 +17,5 @@ __version__: str
 __version_tuple__: VERSION_TUPLE
 version_tuple: VERSION_TUPLE
 
-__version__ = version = '0.25.1'
-__version_tuple__ = version_tuple = (0, 25, 1)
+__version__ = version = '0.26.0'
+__version_tuple__ = version_tuple = (0, 26, 0)
