### 2025-04-28 f10b859
feat: Add workflow to create branch and publish package
Implemented a new GitHub Actions workflow to automate the creation of a
branch and publishing of a package upon release. This includes steps for
checking out the repository, setting up Python, installing dependencies,
building the package, and pushing the branch to the remote repository.

This change enhances the release process by ensuring that packages are
automatically published to PyPI after a release is created.
