+++ .github/workflows/merge-pull-request-to-main.yml
@@ -308,3 +308,50 @@ jobs:
           run: |
             echo "Running post-release scripts..."
             # Add your post-release commands here
+
+  create-branch-and-publish-package:
+    needs: create-github-release
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0  # Fetch all history so we can access all commit logs
+    
+      - name: Set up Python
+        uses: actions/setup-python@v3
+        with:
+          python-version: '3.10'
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install build
+
+      - name: Echo release id
+        run: |
+          echo "RELEASE_ID=${{ github.event.release.id }}"
+          echo "REPO_NAME=${{ github.repository}}"
+          echo "TAG=${{ github.event.release.tag_name }}"
+
+      - name: Build package
+        env:
+            SETUPTOOLS_SCM_PRETEND_VERSION: ${{ github.event.release.tag_name }} 
+        run: python -m build
+
+      - name: Create / update branch
+        run: |
+            git config user.name  "github-actions[bot]"
+            git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
+
+            # Force-create (or fast-forward) the branch to the current commit
+            git checkout -B "${{ github.event.release.tag_name }}"
+            git push --force-with-lease origin "${{ github.event.release.tag_name }}"
+      
+
+      - name: Publish package
+        if: success()
+        uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
+        with:
+          user: __token__
+          password: ${{ secrets.PYPI_API_TOKEN }}
