### 2025-04-28 #7932 @lllangWV
Dear @ckoerber,

Thank you very much for your thorough review of our manuscript and the ParquetDB software. We appreciate the time you took and the constructive feedback provided. We found your comments insightful and helpful in improving the clarity and positioning of our paper.

We have addressed the minor issues you kindly opened directly on the GitHub repository. Below, we address the broader points raised in your review regarding substantial scholarly effort, positioning within the ecosystem, novelty, and usefulness. We have revised the manuscript accordingly, keeping the word count constraints in mind.

# 1. Substantial Scholarly Effort & Positioning within the Ecosystem

> **Concern**: Justifying the "significant contribution" and clarifying ParquetDB's position relative to the Apache stack (Arrow/Parquet), databases (SQLite, MongoDB), and other file formats (HDF5). The reviewer noted ParquetDB seems conceptually closer to file-based storage.

We agree that clearly articulating ParquetDB's specific niche and value proposition is crucial. We have revised the "Statement of Need" section to address this.

- **Comparison with HDF5**: We now explicitly include HDF5 in our discussion of file-based formats, acknowledging its strengths for large numerical datasets but highlighting its limitations as a file container lacking rich, database-like querying APIs.

- **Comparison with Apache Stack (PyArrow/Parquet)**: We have clarified that while ParquetDB builds upon Apache Arrow/Parquet, its contribution lies in the abstraction layer it provides. The revised text explicitly states that ParquetDB offers a familiar database-like interface (CRUD), automates schema management (including evolution across multiple files), simplifies the storage/retrieval of complex Python objects (reducing boilerplate serialization code), and provides a unified API to manage collections of Parquet files as a single logical datastore. We emphasize that achieving this functionality manually with PyArrow requires non-trivial, project-specific code that ParquetDB aims to eliminate.

- **Comparison with Databases (SQLite/MongoDB)**: The comparison has been refined. We acknowledge the ubiquity and strengths of systems like SQLite (serverless, native Python support, SQL capabilities) but better articulate ParquetDB's advantages in specific contexts: superior read performance for wide, columnar data scans, inherent handling of complex/nested data types without ORM or complex JSON handling, schema flexibility, and simple file-based portability without server overhead. We explicitly position ParquetDB as filling the gap where database complexity is overkill, but raw file handling or direct PyArrow usage becomes cumbersome.

- **Conceptual Position**: We acknowledge that ParquetDB lacks features like relational integrity, making it conceptually distinct from traditional RDBMS. However, we maintain that the term "medium ware" accurately reflects its role. By providing a database-like API (CRUD, querying, schema awareness) on top of the file format, it offers significantly more structure and ease of use than direct file manipulation or basic PyArrow Dataset operations, thus bridging the gap identified.

# 2. Assessment of Novelty and Usefulness

> **Concern**: Whether the simplicity gain over PyArrow or SQLite is significant, the performance advantage dependency on workload, and the perceived ease of adding indexing manually via PyArrow.

- **Simplicity**: We refined the "Statement of Need" to emphasize that ParquetDB's simplicity shines particularly when dealing with evolving schemas across multiple files, managing collections of files as a database, and storing complex Python objects directly,  tasks that become increasingly complex using only PyArrow or fitting into SQLite's structure.

- **Performance Dependency** We fully agree that performance benefits are workload-dependent. To address this directly, we added a clarifying sentence to the "Benchmarks" section: "It is important to note that performance advantages depend on the workload; for instance, complex analytical queries involving aggregations or returning small, highly filtered results might favor the mature query engine and indexing of systems like SQLite. ParquetDB excels when querying or returning substantial portions of wide datasets." This explicitly acknowledges the nuance highlighted by the reviewer.

## 3. Further Remarks

### Tests & Converage

> Automated tests are currently not integrated. On my local system, one of the tests fails (see issue).

"Thank you for identifying the lack of CI integration and the local test issue (which has been addressed in the corresponding GitHub issue). We have now implemented automated testing via GitHub Actions. Tests are automatically run using pytest on every pull request to the main branches, ensuring continuous integration.

> Using [pytest](https://docs.pytest.org/en/stable/index.html) and [coverage](https://coverage.readthedocs.io/en/7.8.0/), I measured a test coverage of approximately 73% overall, and 65% for the most relevant module parquetdb/core/parquetdb.py. This seems reasonable but could be improved, particularly for the core functionality (of course the coverage is just a limited predictor for usefulness).

We appreciate you measuring the test coverage. Acknowledging the importance of robust testing, especially for the core functionality, we have dedicated effort to increasing the test coverage to 82% for the parquetdb/core/parquetdb.py. We have added several new tests specifically targeting previously uncovered logic paths within the parquetdb.core module.

### Citations & References

> It may be appropriate to directly cite the following resources in the JOSS paper:
> Apache Software Foundation (Parquet and Arrow projects).
> The longer paper associated with ParquetDB: https://arxiv.org/abs/2502.05311.

Thank you for the suggestions. We have now added citations for the Apache Software Foundation (regarding the Parquet and Arrow projects) and our associated arXiv preprint directly within the JOSS paper manuscript.

### Performance Evaluation

> I was unable to execute the provided benchmarks directly (see issue).

We apologize for the difficulties encountered running the original benchmark scripts (as noted in the relevant GitHub issue). To improve reproducibility, clarity, and accessibility, we have migrated the benchmarks into Jupyter Notebooks. These notebooks now include dedicated sections for specifying hardware configurations and provide integrated discussion alongside the code and results. They are also rendered within the project documentation for easier access.

> Benchmark results should ideally include the hardware specifications (which are provided in the longer paper, but could be summarized here).

Agreed. As mentioned above, the new Jupyter Notebook format for benchmarks explicitly includes sections for documenting the hardware specifications used for each run, addressing this omission

> Based on my own experiments, I find that ParquetDB and SQLite perform similarly for selective "needle-in-a-haystack" queries.
> The performance advantage of ParquetDB appears primarily when reading and returning large datasets, rather than during the filtering step.

We concur with your assessment of the performance characteristics based on your experiments and ours. Our revised manuscript and the new benchmark notebooks now more clearly state that while selective query performance might be comparable to indexed systems like SQLite, ParquetDB's primary performance advantage manifests when reading and returning large volumes of data, owing to the efficiency of the columnar Parquet format and Apache Arrow

> The naive implementation of what ParquetDB does (using a few lines of PyArrow) outperforms ParquetDB in the read/query/write benchmarks

This is an important point, which we also explore in the new benchmark notebooks (specifically the [Small Benchmark.ipynb](https://github.com/lllangWV/ParquetDB/blob/main/examples/benchmarks/Small%20Benchmark.ipynb) derived from your example). We provide a detailed discussion there (summarized in the 'Own Benchmark' section below) explaining that this difference is expected. ParquetDB introduces abstractions (like automated file management, schema handling, indexing/normalization steps) on top of PyArrow, which incur some overhead, particularly during writes, in exchange for significant developer convenience and robustness

### Miscellaneous Suggestions

> The [detailed contributing file](https://lllangwv.github.io/ParquetDB/CONTRIBUTING.html) is not visible in the repository root directory

Acknowledged. The `CONTRIBUTING.md` file has been moved to the repository root directory for better visibility.

> The requirements.txt file may no longer be necessary.

Agreed. As dependencies are managed via `pyproject.toml`, the redundant requirements.txt file has been removed.

> Anaconda environments are listed but not mentioned in the documentation — it might be helpful to clarify this (unless intentionally omitted for licensing reasons).

Good point. To avoid confusion and streamline the repository, the Anaconda environment files, which were not actively documented or maintained, have been removed

> There are GitHub Actions configured for documentation generation and PyPI deployment. I also noticed an interesting feature: an [automated changelog generation script using an OpenAI interface](https://github.com/lllangWV/_ParquetDB_/blob/main/scripts/github_workflow/workflow_generate_changlog_on_recent_tag.py). While this is innovative, I would personally prefer that changelog generation remains a manual (or at least reviewed) step. In contrast, I consider the absence of automated unit tests in the CI workflow more critical.

We agree entirely with your assessment regarding the importance of a review step for the changelog and the higher priority of automated testing (which, as noted above, is now implemented via CI). We appreciate the feedback on the changelog automation. Inspired by workflows like Google's release-please-action, we have redesigned the process. Now, upon merging pull requests, a new pull request is automatically created (or updated) targeting the next release. This 'Release PR' uses an LLM to draft a summary of changes from the merged PRs as a comment. This draft can then be manually reviewed, edited, and finalized within the Release PR before the actual release and tagging occurs, ensuring human oversight while still leveraging automation for the initial draft.

> Some documentation examples could be made more concise by [limiting or summarizing output to improve readability and reduce scrolling](https://lllangwv.github.io/ParquetDB/01_tutorials/03%20-%20Advanced%20Reads.html#Advanced-Read-Operations).

Thank you for this usability suggestion. We have updated the documentation examples, implementing scrollable output blocks for lengthy outputs to improve readability and reduce excessive page scrolling, as suggested.

> The default configuration file is loaded from config = LoggingConfig.from_yaml(os.path.join(UTILS_DIR, "config.yml")). It might be worth considering allowing user-specific configuration files in standard locations (e.g., ~/.config/parquetdb/config.yml on Linux), for which the [platformdirs package](https://pypi.org/project/platformdirs/) could be helpful

Excellent suggestion for enhancing configuration flexibility. We have implemented user-specific configuration file loading using the `platformdirs` package. ParquetDB will now search for config.yml in standard user-specific configuration directories (e.g., ~/.config/parquetdb/ on Linux, etc.) in addition to the package's default location

> The README currently contains the disclaimer:
> "Note: This project is in its initial stages. Features and APIs are subject to change. Please refer to the latest documentation and  release notes for updates."
> If a JOSS publication is targeted, I would suggest that at least the core API should be considered stable.

Agreed. As the core API is stable and we are proceeding with the JOSS submission, the 'initial stages' disclaimer has been removed from the README.

> A useful feature might also be to filter by date of creation (if not already possible)

This functionality is already available. Users can filter data based on datetime fields using the standard filters keyword argument in the read() method, just like filtering on any other data type.

## Own Benchmark

> SQLite exhibits faster write performance and comparable query performance for small result sets.
> ParquetDB significantly outperforms SQLite when reading large datasets (i.e., returning large arrays).
> A naive PyArrow implementation (replicating basic ParquetDB functionality) often outperforms ParquetDB.

Thank you for conducting and sharing your benchmark; we appreciate the independent verification. We have incorporated a version of this benchmark into our repository ([Small Benchmark.ipynb](https://github.com/lllangWV/ParquetDB/blob/main/examples/benchmarks/Small%20Benchmark.ipynb)) and documentation. Our observations align closely with yours, and we summarize the key points and trade-offs as follows:

- **Create/Write Performance**: SQLite indeed demonstrates the fastest raw insert performance for the tested scales, benefiting from its optimized row-based engine. ParquetDB follows, with the naive PyArrow implementation being slower in this specific benchmark setup, potentially due to differences in how data is chunked or indices/metadata are handled during creation compared to ParquetDB's more managed approach. Crucially, ParquetDB's write process includes normalization and indexing steps not present in the naive PyArrow write loop, contributing to this difference.
- **Read Performance (Large Datasets)**: As you noted, both ParquetDB and the naive PyArrow approach significantly outperform SQLite for large reads (full table scans), showcasing the power of columnar storage and zero-copy reads via Apache Arrow. Their performance here is nearly identical, as expected, since ParquetDB leverages Arrow directly for reads.
- **Query Performance (Filtering)**: For vectorized filtering operations on large datasets, ParquetDB and PyArrow again show similar, strong performance, significantly faster than SQLite's row-by-row filtering in Python loops. For highly selective "needle-in-a-haystack" queries where SQLite can leverage its B-tree indexes effectively, performance can be comparable, as noted in the manuscript revision.
- **ParquetDB vs. Naive PyArrow Trade-off**: The observed performance difference (especially in writes) between ParquetDB and a minimal PyArrow script highlights the core value proposition: ParquetDB trades a small performance overhead for significant developer convenience and robustness. It abstracts away manual file/directory management, automatic ID generation, schema evolution handling, consistent indexing/normalization across potentially many underlying files, and provides a simplified CRUD API. This reduces boilerplate code and potential errors, particularly as applications scale or data complexity increases.
- **Key Takeaway**: ParquetDB inherits the excellent read/query performance of Arrow/Parquet for analytical workloads. While SQLite may be faster for write-heavy tasks or simple lookups on indexed fields, ParquetDB offers a compelling balance of performance, features (especially complex type handling and schema flexibility), and developer experience for managing collections of Parquet files as a cohesive datastore, justifying the abstraction layer over using PyArrow directly.

Thank you again for your time and valuable feedback. We found your review very helpful and have worked diligently to address all the points raised.

We've updated the manuscript, codebase, tests, and documentation as described in our detailed responses, and we feel these improvements make ParquetDB and the accompanying paper much stronger.

We hope these revisions meet your expectations and look forward to hearing from you. Please let us know if anything else is required.

Best Regards,

Logan Lang
