### 2025-04-30 e7757a6
feat: Refactor release workflow to streamline branch creation and publishing
Updated the GitHub Actions workflow to improve the process of creating
branches and publishing packages upon release. This includes adding
outputs for release tags and versions, modifying steps to use these
outputs, and removing the deprecated on-release workflow. The changes
ensure a more efficient and clear release process.

Removes the on-release.yml workflow file.
