+++ .github/workflows/merge-pull-request-to-main.yml
@@ -224,6 +224,9 @@ jobs:
     needs: check-pr-status
     if: needs.check-pr-status.outputs.is_release_pr == 'true'
     runs-on: ubuntu-latest
+    outputs:
+      release_tag: ${{ steps.create-tag.outputs.tag_name }}
+      release_version: ${{ steps.extract-version.outputs.version }}
     steps:
         - name: Check out code
           uses: actions/checkout@v3
@@ -262,6 +265,7 @@ jobs:
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
+            pip install openai python-dotenv PyGithub
             pip install .[dev]
 
         - name: Run comment summarization script
@@ -293,65 +297,91 @@ jobs:
             git push origin main
         
         - name: Create and push tag
+          id: create-tag # <<< Added id here
           run: |
-            git tag -a "v${{ steps.extract-version.outputs.version }}" -m "Release v${{ steps.extract-version.outputs.version }}"
-            git push origin "v${{ steps.extract-version.outputs.version }}"
+            TAG_NAME="${{ steps.extract-version.outputs.version }}"
+            echo "Creating tag: $TAG_NAME pointing to merge commit ${{ github.event.pull_request.merge_commit_sha }}"
+            git tag -a "$TAG_NAME" -m "Release $TAG_NAME" "${{ github.event.pull_request.merge_commit_sha }}"
+            git push origin "$TAG_NAME"
+
+            echo "tag_name=$TAG_NAME" >> $GITHUB_OUTPUT # <<< Output the tag name
             
         - name: Create GitHub Release with summarized notes
           run: |
-            # Create GitHub Release using the generated notes
-            gh release create "v${{ steps.extract-version.outputs.version }}" \
-            --title "Release v${{ steps.extract-version.outputs.version }}" \
-            --notes-file RELEASE_NOTES.md
+            echo "Creating release for tag ${{ steps.create-tag.outputs.tag_name }}"
+            gh release create "${{ steps.create-tag.outputs.tag_name }}" \
+              --title "Release ${{ steps.create-tag.outputs.tag_name }}" \
+              --notes-file RELEASE_NOTES.md \
+              --target "${{ github.event.pull_request.merge_commit_sha }}" # Target the merge commit
             
         - name: Run post-release scripts
           run: |
             echo "Running post-release scripts..."
             # Add your post-release commands here
 
-  create-branch-and-publish-package:
-    needs: create-github-release
+  create-release-branch-and-publish:
+    # --- Add dependency ---
+    needs: [create-github-release, check-pr-status] # Depend on release creation AND the check that determined it IS a release PR
+    # --- Add condition --- (Optional but recommended for clarity)
+    # This job should only run if the 'is_release_pr' check was true.
+    # Since it needs create-github-release which *already* has this check,
+    # it will be implicitly skipped if create-github-release is skipped.
+    # Adding the 'if' explicitly makes the intent clearer.
+    if: needs.check-pr-status.outputs.is_release_pr == 'true'
     runs-on: ubuntu-latest
-
     steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0  # Fetch all history so we can access all commit logs
-    
-      - name: Set up Python
-        uses: actions/setup-python@v3
-        with:
-          python-version: '3.10'
+    # --- Checkout code using the tag from the previous job ---
+    - uses: actions/checkout@v4
+      with:
+        fetch-depth: 0  # Fetch all history so we can access the tag
+        # --- Use the output from create-github-release job ---
+        ref: ${{ needs.create-github-release.outputs.release_tag }}
 
-      - name: Install dependencies
[…]
