### 2025-04-30 14eaa10
feat(graph): Introduce graph database structure with nodes and edges
Added new modules for managing graph structures, including NodeStore,
EdgeStore, and ParquetGraphDB. These components facilitate the storage
and retrieval of nodes and edges in a graph format, enhancing the
capabilities of the parquetdb library.

Includes validation functions for node and edge dataframes, as well as
generator stores for managing edge and node generators.

This change lays the foundation for graph-based data management and
operations within the existing parquetdb framework.
