+++ parquetdb/graph/nodes.py
@@ -1,0 +1,427 @@
+import logging
+import os
+from functools import wraps
+from typing import List, Union
+
+import pandas as pd
+import pyarrow as pa
+import pyarrow.compute as pc
+from parquetdb import ParquetDB
+from parquetdb.core.parquetdb import LoadConfig, NormalizeConfig
+
+from parquetdb.graph.generator_store import validate_generator_inputs
+from parquetdb.graph.utils import get_dataframe_column_names
+
+logger = logging.getLogger(__name__)
+
+
+REQUIRED_NODE_COLUMNS_FIELDS = set()
+
+
+def validate_node_dataframe(df):
+    column_names = get_dataframe_column_names(df)
+    fields = set(column_names)
+    missing_fields = REQUIRED_NODE_COLUMNS_FIELDS - fields
+    if missing_fields:
+        raise ValueError(
+            f"Node dataframe is missing required fields: {missing_fields}. Node dataframe must contain the following columns: {REQUIRED_NODE_COLUMNS_FIELDS}"
+        )
+    return df
+
+
+def node_generator(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        # Perform pre-execution checks
+        logger.debug(f"Executing {func.__name__} with args: {args}, kwargs: {kwargs}")
+        validate_generator_inputs(args)
+
+        df = func(*args, **kwargs)
+        validate_node_dataframe(df)
+        return df
+
+    return wrapper
+
+
+class NodeStore(ParquetDB):
+    """
+    A wrapper around ParquetDB specifically for storing node features
+    of a given node type.
+    """
+
+    node_metadata_keys = ["class", "class_module", "node_type", "name_column"]
+
+    def __init__(self, storage_path: str, initialize_kwargs: dict = None):
+        """
+        Parameters
+        ----------
+        storage_path : str
+            The path where ParquetDB files for this node type are stored.
+        """
+
+        self.node_type = os.path.basename(storage_path)
+
+        if initialize_kwargs is None:
+            initialize_kwargs = {}
+
+        super().__init__(db_path=storage_path)
+
+        metadata = self.get_metadata()
+
+        update_metadata = False
+        for key in self.node_metadata_keys:
+            if key not in metadata:
+                update_metadata = update_metadata or key not in metadata
+        if update_metadata:
+            self.set_metadata(
+                {
+                    "class": f"{self.__class__.__name__}",
+                    "class_module": f"{self.__class__.__module__}",
+                    "node_type": self.node_type,
+                    "name_column": "id",
+                }
+            )
+
+        if self.is_empty():
+            self._initialize(**initialize_kwargs)
+
+        logger.debug(f"Initialized NodeStore at {storage_path}")
+
+    def __repr__(self):
+        return self.summary(show_column_names=False)
+
+    @property
+    def storage_path(self):
+        return self._db_path
+
+    @storage_path.setter
+    def storage_path(self, value):
+        self._db_path = value
+        self.node_type = os.path.basename(value)
+
+    def _initialize(self, **kwargs):
+        data = self.initialize(**kwargs)
+        if data is not None:
+            self.create_nodes(data=data)
+
+    def initialize(self, **kwargs):
+        return None
+
+    @property
+    def name_column(self):
+        return self.get_metadata()["name_column"]
+
+    @name_column.setter
+    def name_column(self, value):
+        self.set_metadata({"name_column": value})
+
+    @property
+    def n_nodes(self):
+        return self.read_nodes(columns=["id"]).num_rows
+
+    @property
+    def n_features(self):
+        return len(self.get_schema().names)
+
+    @property
+    def columns(self):
[…]
