+++ parquetdb/graph/edges.py
@@ -1,0 +1,441 @@
+import logging
+import os
+from functools import wraps
+from typing import Dict, List, Union
+
+import pandas as pd
+import pyarrow as pa
+import pyarrow.compute as pc
+from parquetdb import ParquetDB
+from parquetdb.core.parquetdb import LoadConfig, NormalizeConfig
+
+from parquetdb.graph.utils import get_dataframe_column_names
+
+logger = logging.getLogger(__name__)
+
+REQUIRED_EDGE_COLUMNS_FIELDS = set(
+    ["source_id", "source_type", "target_id", "target_type", "edge_type"]
+)
+
+
+def validate_edge_dataframe(df):
+    column_names = get_dataframe_column_names(df)
+    fields = set(column_names)
+    missing_fields = REQUIRED_EDGE_COLUMNS_FIELDS - fields
+    if missing_fields:
+        raise ValueError(
+            f"Edge dataframe is missing required fields: {missing_fields}. Edge dataframe must contain the following columns: {REQUIRED_EDGE_COLUMNS_FIELDS}"
+        )
+    return df
+
+
+def edge_generator(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        # Perform pre-execution checks
+        logger.debug(f"Executing {func.__name__} with args: {args}, kwargs: {kwargs}")
+        df = func(*args, **kwargs)
+        validate_edge_dataframe(df)
+        return df
+
+    wrapper.__name__ = func.__name__
+    wrapper.__doc__ = func.__doc__
+    return wrapper
+
+
+class EdgeStore(ParquetDB):
+    """
+    A wrapper around ParquetDB specifically for storing edge features
+    of a given edge type.
+    """
+
+    required_fields = REQUIRED_EDGE_COLUMNS_FIELDS
+    edge_metadata_keys = ["class", "class_module"]
+
+    def __init__(self, storage_path: str, setup_kwargs: dict = None):
+        """
+        Parameters
+        ----------
+        storage_path : str
+            The path where ParquetDB files for this edge type are stored.
+        """
+
+        super().__init__(
+            db_path=storage_path,
+            initial_fields=[
+                pa.field("source_id", pa.int64()),
+                pa.field("source_type", pa.string()),
+                pa.field("target_id", pa.int64()),
+                pa.field("target_type", pa.string()),
+                pa.field("edge_type", pa.string()),
+            ],
+        )
+
+        self._initialize_metadata()
+        self._initialize_field_metadata()
+
+        logger.debug(f"Initialized EdgeStore at {storage_path}")
+        if self.is_empty():
+            if setup_kwargs is None:
+                setup_kwargs = {}
+            self._setup(**setup_kwargs)
+
+    def __repr__(self):
+        return self.summary(show_column_names=False)
+
+    @property
+    def storage_path(self):
+        return self._db_path
+
+    @storage_path.setter
+    def storage_path(self, value):
+        self._db_path = value
+        self.edge_type = os.path.basename(value)
+
+    @property
+    def edge_type(self):
+        return os.path.basename(self.storage_path)
+
+    @edge_type.setter
+    def edge_type(self, value):
+        self._edge_type = value
+
+    @property
+    def n_edges(self):
+        return self.read_edges(columns=["id"]).num_rows
+
+    @property
+    def n_features(self):
+        return len(self.get_schema().names)
+
+    @property
+    def columns(self):
+        return self.get_schema().names
+
+    def summary(self, show_column_names: bool = False):
+        fields_metadata = self.get_field_metadata()
+        metadata = self.get_metadata()
[…]
