### 2025-04-30 a87a8f8
feat(tests): Add comprehensive tests for graph database components
Introduced unit tests for EdgeStore, NodeStore, and GraphDB to ensure
correct functionality of graph database operations. Tests cover
initialization, edge and node creation, updates, deletions, and
filtering, as well as generator functionality.

These tests enhance the reliability of the graph database features
within the parquetdb library and facilitate future development and
maintenance.
