+++ tests/graph/test_parquet_graphdb.py
@@ -1,0 +1,613 @@
+import os
+import shutil
+
+import pandas as pd
+import pyarrow as pa
+import pytest
+
+from parquetdb.graph import EdgeStore, GraphDB, NodeStore
+from parquetdb.utils import element, element_element_neighborsByGroupPeriod
+from parquetdb.utils.config import PKG_DIR, config
+
+config.logging_config.loggers.matgraphdb.level = "DEBUG"
+
+config.apply()
+
+
+@pytest.fixture
+def tmp_dir(tmp_path):
+    """Fixture for temporary directory."""
+    tmp_dir = str(tmp_path)
+    yield tmp_dir
+    if os.path.exists(tmp_dir):
+        shutil.rmtree(tmp_dir)
+
+
+@pytest.fixture
+def graphdb(tmp_dir):
+    """Fixture to create a GraphDB instance."""
+    return GraphDB(storage_path=tmp_dir)
+
+
+@pytest.fixture
+def element_store(tmp_dir):
+    """Fixture to create an ElementNodes instance."""
+    element_store = NodeStore(storage_path=os.path.join(tmp_dir, "elements"))
+    element_store.create_nodes(elements())
+    return element_store
+
+
+@pytest.fixture
+def test_data():
+    nodes_1_data = [{"name": "Source1"}, {"name": "Source2"}]
+    nodes_2_data = [{"name": "Target1"}, {"name": "Target2"}]
+    node_1_type = "user"
+    node_2_type = "item"
+    edge_data = [
+        {
+            "source_id": 0,
+            "source_type": node_1_type,
+            "target_id": 0,
+            "target_type": node_2_type,
+            "weight": 0.5,
+        },
+        {
+            "source_id": 1,
+            "source_type": node_2_type,
+            "target_id": 1,
+            "target_type": node_1_type,
+            "weight": 0.7,
+        },
+        {
+            "source_id": 0,
+            "source_type": node_1_type,
+            "target_id": 1,
+            "target_type": node_2_type,
+            "weight": 0.5,
+        },
+        {
+            "source_id": 1,
+            "source_type": node_2_type,
+            "target_id": 0,
+            "target_type": node_1_type,
+            "weight": 0.7,
+        },
+    ]
+    edge_type = "test_edge"
+    return nodes_1_data, node_1_type, nodes_2_data, node_2_type, edge_data, edge_type
+
+
+@pytest.fixture
+def wyckoff_generator():
+    """Fixture to provide a sample node generator function."""
+
+    def generate_wyckoff_nodes():
+        """Generate basic Wyckoff position nodes."""
+        data = [
+            {"symbol": "1a", "multiplicity": 1, "letter": "a", "site_symmetry": "1"},
+            {"symbol": "2b", "multiplicity": 2, "letter": "b", "site_symmetry": "2"},
+        ]
+        return pa.Table.from_pylist(data)
+
+    return generate_wyckoff_nodes
+
+
+@pytest.fixture
+def node_generator_data(wyckoff_generator):
+    """Fixture providing test data for node generators."""
+    generator_name = "test_wyckoff_generator"
+    generator_func = wyckoff_generator
+    generator_args = {}
+    generator_kwargs = {}
+    return generator_name, generator_func, generator_args, generator_kwargs
+
+
+def test_initialize_graphdb(graphdb):
+    """Test if GraphDB initializes with the correct directories."""
+    assert os.path.exists(graphdb.nodes_path), "Nodes directory not created."
+    assert os.path.exists(graphdb.edges_path), "Edges directory not created."
+    assert os.path.exists(
+        graphdb.edge_generators_path
+    ), "Edge generators directory not created."
+    assert os.path.exists(graphdb.graph_path), "Graph directory not created."
+
+
[…]
