+++ tests/graph/test_edge_store.py
@@ -1,0 +1,177 @@
+import os
+import shutil
+
+import pandas as pd
+import pyarrow as pa
+import pytest
+
+from parquetdb.graph import EdgeStore, NodeStore
+
+from utils import element
+
+
+@pytest.fixture
+def temp_storage(tmp_path):
+    """Fixture to create and cleanup a temporary storage directory"""
+    storage_dir = tmp_path / "test_edge_store"
+    yield str(storage_dir)
+    if os.path.exists(storage_dir):
+        shutil.rmtree(storage_dir)
+
+
+@pytest.fixture
+def tmp_dir(tmp_path):
+    """Fixture for temporary directory."""
+    tmp_dir = str(tmp_path)
+    yield tmp_dir
+    if os.path.exists(tmp_dir):
+        shutil.rmtree(tmp_dir)
+
+
+@pytest.fixture
+def edge_store(tmp_dir):
+    """Fixture to create an EdgeStore instance."""
+    edge_store = EdgeStore(storage_path=os.path.join(tmp_dir, "edges"))
+    return edge_store
+
+
+@pytest.fixture
+def element_store(tmp_dir):
+    """Fixture to create an ElementNodes instance."""
+    element_store = NodeStore(storage_path=os.path.join(tmp_dir, "elements"))
+    element_store.create_nodes(element())
+    return element_store
+
+
+@pytest.fixture
+def sample_edge_data():
+    """Fixture providing sample edge data with required fields"""
+    return {
+        "source_id": [1, 2],
+        "target_id": [3, 4],
+        "source_type": ["node_a", "node_a"],
+        "target_type": ["node_b", "node_b"],
+        "edge_type": ["has", "has"],
+        "weight": [0.5, 0.7],
+    }
+
+
+def test_edge_store_initialization(temp_storage):
+    """Test that EdgeStore initializes correctly and creates the storage directory"""
+    store = EdgeStore(temp_storage)
+    assert os.path.exists(temp_storage)
+    assert store is not None
+
+    metadata = store.get_metadata()
+    assert metadata["class"] == "EdgeStore"
+    assert metadata["class_module"] == "matgraphdb.core.edges"
+
+
+def test_create_edges_from_dict(edge_store, sample_edge_data):
+    """Test creating edges from a dictionary"""
+    edge_store.create_edges(sample_edge_data)
+
+    # Read back and verify
+    result_table = edge_store.read_edges()
+    result_df = result_table.to_pandas()
+    assert len(result_df) == 2
+    assert all(field in result_df.columns for field in EdgeStore.required_fields)
+    assert list(result_df["source_id"]) == [1, 2]
+    assert list(result_df["target_id"]) == [3, 4]
+
+
+def test_create_edges_from_dataframe(edge_store, sample_edge_data):
+    """Test creating edges from a pandas DataFrame"""
+    df = pd.DataFrame(sample_edge_data)
+    edge_store.create_edges(df)
+
+    result_table = edge_store.read_edges()
+    result_df = result_table.to_pandas()
+    assert len(result_df) == 2
+    assert all(result_df["source_id"] == df["source_id"])
+    assert all(result_df["target_id"] == df["target_id"])
+
+
+def test_read_edges_with_filters(edge_store, sample_edge_data):
+    """Test reading edges with specific filters"""
+    edge_store.create_edges(sample_edge_data)
+
+    # Read with column filter
+    result_table = edge_store.read_edges(columns=["id", "source_id", "target_id"])
+    result_df = result_table.to_pandas()
+    assert list(result_df.columns) == ["id", "source_id", "target_id"]
+
+    # Read with ID filter
+    first_result_table = edge_store.read_edges()
[…]
