### 2025-04-30 ec02b44
refactor(tests): Replace GraphDB with ParquetGraphDB in test suite
Updated test cases to utilize ParquetGraphDB instead of GraphDB for
consistency with recent architectural changes. This refactor ensures
that all tests are aligned with the new graph database implementation
and maintains the integrity of the testing framework.

No changes to test logic were made, only the underlying database
class was updated.
