+++ tests/graph/test_parquet_graphdb.py
@@ -1,17 +1,15 @@
 import os
 import shutil
-
+import logging
 import pandas as pd
 import pyarrow as pa
 import pytest
 
-from parquetdb.graph import EdgeStore, GraphDB, NodeStore
-from parquetdb.utils import element, element_element_neighborsByGroupPeriod
-from parquetdb.utils.config import PKG_DIR, config
+from parquetdb.graph import EdgeStore, ParquetGraphDB, NodeStore
 
-config.logging_config.loggers.matgraphdb.level = "DEBUG"
+from utils import element, element_element_neighborsByGroupPeriod
 
-config.apply()
+logger = logging.getLogger(__name__)
 
 
 @pytest.fixture
@@ -26,14 +24,14 @@ def tmp_dir(tmp_path):
 @pytest.fixture
 def graphdb(tmp_dir):
     """Fixture to create a GraphDB instance."""
-    return GraphDB(storage_path=tmp_dir)
+    return ParquetGraphDB(storage_path=tmp_dir)
 
 
 @pytest.fixture
 def element_store(tmp_dir):
     """Fixture to create an ElementNodes instance."""
     element_store = NodeStore(storage_path=os.path.join(tmp_dir, "elements"))
-    element_store.create_nodes(elements())
+    element_store.create_nodes(element())
     return element_store
 
 
@@ -169,7 +167,7 @@ def test_add_node_store(tmp_dir):
     temp_store.create_nodes(test_data)
 
     # Create a graph database
-    graph = GraphDB(storage_path=os.path.join(tmp_dir, "graph"))
+    graph = ParquetGraphDB(storage_path=os.path.join(tmp_dir, "graph"))
 
     # Add the node store to the graph
     graph.add_node_store(temp_store)
@@ -211,7 +209,7 @@ def test_add_node_store_with_remove_original(tmp_dir):
     temp_store.normalize()  # Add this line to ensure data is written
 
     # Create a graph database
-    graphdb = GraphDB(storage_path=os.path.join(tmp_dir, "graph"))
+    graphdb = ParquetGraphDB(storage_path=os.path.join(tmp_dir, "graph"))
 
     # Add the node store with remove_original=True
     graphdb.add_node_store(temp_store, remove_original=True)
@@ -233,7 +231,7 @@ def test_add_node_store_with_remove_original(tmp_dir):
 def test_nodes_persist_after_reload(tmp_dir):
     """Test that nodes persist and can be loaded after recreating the GraphDB instance."""
     # Create initial graph instance and add nodes
-    graph = GraphDB(storage_path=tmp_dir, load_custom_stores=False)
+    graph = ParquetGraphDB(storage_path=tmp_dir, load_custom_stores=False)
     node_type = "test_node"
     test_data = [{"name": "Node1", "value": 10}, {"name": "Node2", "value": 20}]
     graph.add_nodes(node_type, test_data)
@@ -247,7 +245,7 @@ def test_nodes_persist_after_reload(tmp_dir):
     ], "Incorrect node names."
 
     # Create new graph instance (simulating program restart)
-    new_graph = GraphDB(storage_path=tmp_dir, load_custom_stores=False)
+    new_graph = ParquetGraphDB(storage_path=tmp_dir, load_custom_stores=False)
 
     # Verify data persisted
     loaded_nodes = new_graph.get_nodes(node_type)
@@ -385,7 +383,7 @@ def test_remove_edge_store(graphdb, test_data):
 def test_edges_persist_after_reload(tmp_dir, test_data):
     """Test that edges persist and can be loaded after recreating the GraphDB instance."""
     # Create initial graph instance and add edges
-    graph = GraphDB(storage_path=tmp_dir)
+    graph = ParquetGraphDB(storage_path=tmp_dir)
[…]
