### 2025-04-30 17db7e2
feat(graphdb): Add ParquetGraphDB examples and notebooks for management
Introduced new example notebooks for managing graphs in MatGraphDB,
including functionalities for creating and managing node and edge
types. Updated the index.rst to include links to the new GraphDB
gallery and added relevant content to showcase the capabilities of
the graph database.

These additions enhance the documentation and provide practical
examples for users to understand and utilize the graph management
features effectively.
