### 2025-04-30 e5ce684
feat(graphdb): Update example notebooks and remove obsolete file
Updated the index.rst to reflect changes in example notebooks for
ParquetGraphDB, including the removal of the obsolete
"02 - Graph Generators in ParquetGraphDB.ipynb" file. This
cleanup ensures that the documentation remains relevant and
focused on current functionalities.

The changes improve the clarity and usability of the example
notebooks for users.
