### 2025-04-30 2109100
bug(tests): bug in edge test in ParquetGraphDB. ed
Updated test cases for ParquetGraphDB for edges. "edge_type" is required for input.
