+++ tests/graph/test_parquet_graphdb.py
@@ -1,15 +1,18 @@
+import logging
 import os
 import shutil
-import logging
+
 import pandas as pd
 import pyarrow as pa
 import pytest
-
-from parquetdb.graph import EdgeStore, ParquetGraphDB, NodeStore
-
 from utils import element, element_element_neighborsByGroupPeriod
 
+from parquetdb.graph import EdgeStore, NodeStore, ParquetGraphDB
+from parquetdb.utils.log_utils import set_verbose_level
+
 logger = logging.getLogger(__name__)
+
+VERBOSE = 2
 
 
 @pytest.fixture
@@ -24,13 +27,15 @@ def tmp_dir(tmp_path):
 @pytest.fixture
 def graphdb(tmp_dir):
     """Fixture to create a GraphDB instance."""
-    return ParquetGraphDB(storage_path=tmp_dir)
+    return ParquetGraphDB(storage_path=tmp_dir, verbose=VERBOSE)
 
 
 @pytest.fixture
 def element_store(tmp_dir):
     """Fixture to create an ElementNodes instance."""
-    element_store = NodeStore(storage_path=os.path.join(tmp_dir, "elements"))
+    element_store = NodeStore(
+        storage_path=os.path.join(tmp_dir, "elements"), verbose=VERBOSE
+    )
     element_store.create_nodes(element())
     return element_store
 
@@ -41,12 +46,14 @@ def test_data():
     nodes_2_data = [{"name": "Target1"}, {"name": "Target2"}]
     node_1_type = "user"
     node_2_type = "item"
+    edge_type = "test_edge"
     edge_data = [
         {
             "source_id": 0,
             "source_type": node_1_type,
             "target_id": 0,
             "target_type": node_2_type,
+            "edge_type": edge_type,
             "weight": 0.5,
         },
         {
@@ -54,6 +61,7 @@ def test_data():
             "source_type": node_2_type,
             "target_id": 1,
             "target_type": node_1_type,
+            "edge_type": edge_type,
             "weight": 0.7,
         },
         {
@@ -61,6 +69,7 @@ def test_data():
             "source_type": node_1_type,
             "target_id": 1,
             "target_type": node_2_type,
+            "edge_type": edge_type,
             "weight": 0.5,
         },
         {
@@ -68,10 +77,10 @@ def test_data():
             "source_type": node_2_type,
             "target_id": 0,
             "target_type": node_1_type,
+            "edge_type": edge_type,
             "weight": 0.7,
         },
     ]
-    edge_type = "test_edge"
     return nodes_1_data, node_1_type, nodes_2_data, node_2_type, edge_data, edge_type
 
 
@@ -160,14 +169,14 @@ def test_add_node_store(tmp_dir):
     """Test adding a node store to the graph database."""
     # Create a temporary node store
     temp_store_path = os.path.join(tmp_dir, "temp_store")
-    temp_store = NodeStore(storage_path=temp_store_path)
+    temp_store = NodeStore(storage_path=temp_store_path, verbose=VERBOSE)
 
     # Add some test data to the temp store
     test_data = [{"name": "Node1"}, {"name": "Node2"}]
     temp_store.create_nodes(test_data)
 
     # Create a graph database
-    graph = ParquetGraphDB(storage_path=os.path.join(tmp_dir, "graph"))
+    graph = ParquetGraphDB(storage_path=os.path.join(tmp_dir, "graph"), verbose=VERBOSE)
 
     # Add the node store to the graph
     graph.add_node_store(temp_store)
@@ -177,7 +186,9 @@ def test_add_node_store(tmp_dir):
     assert len(graph.get_nodes(temp_store.node_type)) == 2
 
[…]
