### 2025-04-30 a2132e3
feat(graph): Add verbose parameter to graph store initializations
Introduced a verbose parameter to the constructors of EdgeStore,
GeneratorStore, NodeStore, and ParquetGraphDB to enhance logging
capabilities. This allows users to control the verbosity of log
messages during initialization and operations, improving debugging
and monitoring of the graph database components.

This change ensures better visibility into the internal processes
of the graph database.
