+++ parquetdb/graph/generator_store.py
@@ -7,10 +7,12 @@ from typing import Callable, Dict, List, Union
 import pandas as pd
 import pyarrow as pa
 import pyarrow.compute as pc
+
 from parquetdb import ParquetDB
 from parquetdb.core import types
 from parquetdb.core.parquetdb import NormalizeConfig
 from parquetdb.utils import data_utils
+from parquetdb.utils.log_utils import set_verbose_level
 
 logger = logging.getLogger(__name__)
 
@@ -38,7 +40,12 @@ class GeneratorStore(ParquetDB):
     required_fields = ["generator_name", "generator_func"]
     metadata_keys = ["class", "class_module"]
 
-    def __init__(self, storage_path: str, initial_fields: List[pa.Field] = None):
+    def __init__(
+        self,
+        storage_path: str,
+        initial_fields: List[pa.Field] = None,
+        verbose: int = 1,
+    ):
         """
         Initialize the EdgeGeneratorStore.
 
@@ -48,6 +55,7 @@ class GeneratorStore(ParquetDB):
             Path where the generator functions will be stored
 
         """
+
         if initial_fields is None:
             initial_fields = []
 
@@ -57,7 +65,12 @@ class GeneratorStore(ParquetDB):
                 pa.field("generator_func", types.PythonObjectArrowType()),
             ]
         )
-        super().__init__(db_path=storage_path, initial_fields=initial_fields)
+        super().__init__(
+            db_path=storage_path,
+            initial_fields=initial_fields,
+            serialize_python_objects=True,
+            verbose=verbose,
+        )
 
         self._initialize_metadata()
         logger.debug(f"Initialized GeneratorStore at {storage_path}")
