### 2025-04-30 9560b6f
feat(tests): Implement edge and node generator tests for GraphDB
Added tests for adding and running edge and node generators in the
GraphDB. The tests verify the correct addition of generators, their
output structure, and persistence across reloads. This ensures that
the generator functionality works as expected and maintains data
integrity.

Included validation checks for output tables and error handling for
invalid generator arguments to enhance test coverage and reliability.
