+++ tests/graph/test_parquet_graphdb.py
@@ -429,210 +429,211 @@ def test_edges_persist_after_reload(tmp_dir, test_data):
     assert edge_dict["weight"] == [0.5, 0.7, 0.5, 0.7], "Incorrect edge weights."
 
 
-# def test_add_edge_generator(graphdb, element_store):
-#     """Test adding an edge generator to the GraphDB."""
-#     # Get the generator name from the function
+def test_add_edge_generator(graphdb, element_store):
+    """Test adding an edge generator to the GraphDB."""
+    # Get the generator name from the function
 
-#     generator_name = element_element_neighborsByGroupPeriod.__name__
+    generator_name = element_element_neighborsByGroupPeriod.__name__
 
-#     graphdb.add_node_store(element_store)
-#     # Add the generator
-#     graphdb.add_edge_generator(
-#         element_element_neighborsByGroupPeriod,
-#         generator_args={"element_store": element_store},
-#     )
+    graphdb.add_node_store(element_store)
+    # Add the generator
+    graphdb.add_edge_generator(
+        element_element_neighborsByGroupPeriod,
+        generator_args={"element_store": element_store},
+    )
 
-#     # Verify the generator was added
-#     assert graphdb.edge_generator_store.is_in(generator_name)
-
-
-# def test_run_edge_generator(graphdb, element_store):
-#     """Test running an edge generator and verify its output."""
-
-#     graphdb.add_node_store(element_store)
-
-#     generator_name = element_element_neighborsByGroupPeriod.__name__
-
-#     # Add and run the generator
-#     graphdb.add_edge_generator(
-#         element_element_neighborsByGroupPeriod,
-#         generator_args={"element_store": element_store},
-#     )
-
-#     table = graphdb.run_edge_generator(generator_name)
-
-#     # Verify the output table has the expected structure
-#     assert isinstance(table, pa.Table)
-#     expected_columns = {
-#         "source_id",
-#         "target_id",
-#         "source_type",
-#         "target_type",
-#         "weight",
-#         "source_name",
-#         "target_name",
-#         "name",
-#         "source_extended_group",
-#         "source_period",
-#         "target_extended_group",
-#         "target_period",
-#     }
-#     assert set(table.column_names) == expected_columns
-
-#     # Convert to pandas for easier verification
-#     df = table.to_pandas()
-
-#     # Basic validation checks
-#     assert not df.empty, "Generator produced no edges"
-#     assert all(df["source_type"] == "elements"), "Incorrect source_type"
-#     assert all(df["target_type"] == "elements"), "Incorrect target_type"
-#     assert all(df["weight"] == 1.0), "Incorrect weight values"
-
-#     # Verify edge names are properly formatted
-#     assert all(
-#         df["name"].str.contains("_neighborsByGroupPeriod_")
-#     ), "Edge names not properly formatted"
+    # Verify the generator was added
+    assert graphdb.edge_generator_store.is_in(generator_name)
 
 
-# def test_edge_generator_persistence(tmp_dir, element_store):
-#     """Test that edge generators persist when reloading the GraphDB."""
-#     generator_name = element_element_neighborsByGroupPeriod.__name__
+def test_run_edge_generator(graphdb, element_store):
+    """Test running an edge generator and verify its output."""
 
-#     # Create initial graph instance and add generator
-#     graph = ParquetGraphDB(storage_path=tmp_dir, verbose=VERBOSE)
-#     graph.add_node_store(element_store)
-#     graph.add_edge_generator(
-#         element_element_neighborsByGroupPeriod,
-#         generator_args={"element_store": element_store},
-#     )
[…]
