### 2025-04-30 bc5f57c
fix(tests): Suppress DeprecationWarning in ParquetGraphDB tests
Added a filter to ignore DeprecationWarning related to the __array__
implementation in the test suite for ParquetGraphDB. This change
prevents unnecessary warning messages during test execution,
allowing for cleaner test output and improved focus on relevant
test results.

This adjustment ensures that the test environment remains stable
and user-friendly.
