### 2025-05-04 lllangWV/ParquetDB#10 @ckoerber
Hello @lllangWV

this solves the issue. Thanks!
