### 2025-05-04 lllangWV/ParquetDB#9 @ckoerber
Hello @lllangWV

Looks good. Thank you!
