### 2025-05-04 lllangWV/ParquetDB#8 @ckoerber
Hello @lllangWV

indeed, with the most recent release (which forces pyarrow==17.0.0),  the tests run again. Hence, this temporary fix solves this issue. Thanks!
