### 2025-05-04 #7932 @ckoerber
Dear @lllangWV ,

Thank you for your detailed response and the updates to the repository and documentation. All of my comments have been fully addressed. @fabian-s, I support the publication of this work in JOSS.
