### 2025-05-05 lllangWV/ParquetDB#29 @perdelt
Fail to install
When I try to update the package with `pip install parquetdb --upgrade` it fails with `Failed to build pyarrow`. This may be related to the shift to pyproject.toml (?)

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7932
