### 2025-05-05 lllangWV/ParquetDB#29 @lllangWV
Hey,

I believe this is an issue between the python version and the pyarrow version. I've noticed pyarrow can't build with `python==3.13`. To fix this, I have updated the pyproject.toml to require python version less than `3.13`.

This might be related to Issue #8 where I have to limit the the version of `pyarrow==17.0.0` as a temporary fix. The issue is coming from pandas, which should be fixed when they release `pandas==3.0`, until then I have to keep to `pyarrow==17.0.0`.

At least that is my guess at what the issue is. Can you give me more details of your environment?

What python version do you have?
What pyarrow version do you have installed?

Logan Lang
