### 2025-05-05 lllangWV/ParquetDB#29 @perdelt
Yes, I have `Python 3.13.2` and it's `pyarrow==19.0.1`. This also happens when I create a new conda environment.

I now have also tried a new environment with `Python 3.12.9`. Everything works fine then. Thanks!

Maybe you could add a remark about that somewhere or patch the installer?
