### 2025-05-05 lllangWV/ParquetDB#31 @perdelt
Link to CONTRIBUTING.md
I think it would help seeing the CONTRIBUTING.md if it would be linked directly from the "Contributing" section in the main README

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7932
