### 2025-05-05 0365c98
refactor(graphdb): Change return type of custom node-generation function
Updated the return type of the custom node-generation function in
ParquetGraphDB from None to pa.Table. This change clarifies the
expected output of the function, improving type safety and
documentation for users implementing custom generators.
