### 2025-05-05 7cbbc40
refactor(tests): Update graphdb fixture to include GraphDB directory
Modified the graphdb fixture in the test_parquet_graphdb.py file to
create a specific GraphDB directory within the temporary directory.
This change improves the organization of test data and ensures that
the GraphDB instance is correctly initialized with a dedicated storage
path, enhancing test reliability.
