### 2025-05-05 c7bde76
refactor(graphdb): Comment out debug log for node stores
Removed the debug log statement for node stores in the node_exists
method to reduce log verbosity. This change aims to streamline logging
output while maintaining the functionality of the method.
