### 2025-05-05 6277838
refactor(graphdb): Update storage path assignment in ParquetGraphDB
Changed the assignment of storage_path in ParquetGraphDB to use the
raw storage_path instead of its absolute path. This adjustment
simplifies the initialization process and maintains the original
directory structure as intended by the user.

This change aims to improve clarity and usability when setting
up the GraphDB instance.
