### 2025-05-05 9128aa5
refactor(graphdb): Update storage path retrieval to use absolute paths
Changed the retrieval of storage paths in EdgeStore, NodeStore, and
ParquetGraphDB to use absolute paths instead of relative paths. This
modification ensures consistency in path representation and improves
clarity when accessing storage locations.

This change aims to enhance usability and reduce potential confusion
related to path handling in the graph database.
