### 2025-05-06 lllangWV/ParquetDB#29 @lllangWV
Hey @perdelt,

Thanks for confirming that using Python 3.12.9 resolved the `pyarrow` build issue on your end, and for your suggestion to add a remark or patch the installer!

I've now updated the `pyproject.toml` to require Python versions between 3.8 and 3.12 (inclusive) and to pin `pyarrow==17.0.0`. This measure addresses both:
1.  The build failures of `pyarrow` when using Python 3.13+.
2.  The test failures detailed in Issue #8, which occur with `pyarrow` versions 18.0.0 and newer.

The core of Issue #8, as you've seen, is an `AssertionError: DataType(binary) != PythonObjectArrowType(DataType(binary))` in `test_arrow_conversion` when using newer `pyarrow` versions. This test validates a specific feature of ParquetDB: the ability to seamlessly store and retrieve arbitrary Python objects (like class instances or Python functions). This is achieved by:
- Detecting fields with Python objects.
- Creating custom `pyarrow` and `pandas` extension types.
- These types automatically serialize the Python objects using `dill` for storage (as binary in Arrow tables) and deserialize them back into Python objects when accessed in a Pandas DataFrame.

While I believe this is a convenient feature for handling complex data types directly, the current situation means that to keep it functional, ParquetDB is restricted to older versions of `pyarrow` (and consequently, cannot easily support Python 3.13+ due to `pyarrow`'s own build constraints with that Python version). This creates a trade-off between retaining this specific object-storage functionality and allowing users to benefit from the latest versions of `pyarrow` and Python.

I am actively looking into how this custom type handling can be adapted to work correctly with newer `pyarrow` versions to resolve Issue #8 more permanently and lift these version restrictions.

In the meantime, given your role in the JOSS review, I would greatly appreciate your perspective on this:

- How valuable or critical do you consider the feature of being able to directly store and retrieve arbitrary Python objects (like class instances or functions) within ParquetDB?
- From a user's or reviewer's standpoint, would it be preferable to:
    - Keep the current version restrictions on Python and PyArrow to ensure this Python object storage feature remains stable?
    - Prioritize compatibility with newer Python and PyArrow versions, even if it means temporarily altering or removing this specific feature until a fully compatible solution is developed?

Thanks again for your help and feedback!

Best,
Logan Lang
