### 2025-05-06 lllangWV/ParquetDB#29 @perdelt
Hey @lllangWV

Thank you for the very detailed analysis and the proposed solution.

I think the current implementation is best. If an important feature doesn't work because a version of a required package contains a bug, you have to discard that version (and wait for a newer, working version). I think that makes a lot of sense.

Compatibility with Python structures seems important to me. I don't know if it's critical, but I would consider it very valuable, so stick with your initial intuition would be my advice.
