### 2025-05-06 #7932 @lllangWV
Dear [perdelt](https://github.com/perdelt),

Thank you for your feedback and for taking the time to thoroughly evaluate ParquetDB. We appreciate your positive comments on the documentation and scholarly effort, as well as your recognition of the recent improvements to the package and paper.

We are happy to address your points regarding "normalization" and the "statement of need."

### On the term "Normalization"

We understand your point that the term "normalization" can be misleading, especially in a database context where it often refers to database schema normalization (e.g., 1NF, 2NF, 3NF). In ParquetDB, this term has been used with a different intention, serving a dual role:

1.  **Optimizing Parquet File Structure:** Firstly, it refers to the process of adjusting the number of row-groups and rows per Parquet file to align with recommended optimal sizes (approximately 200MB per row group and around 2GB per Parquet file) for better performance.
2.  **Summarizing Data Preprocessing Contributions:** Secondly, the term "data normalization" was suggested by a previous reviewer to encapsulate the key contributions of ParquetDB beyond the basic Apache Parquet/Arrow functionalities. As they noted, this includes "functionality to merge, check, and clean multiple Parquet files into consistent datasets."

Perhaps using "data normalization" is too imprecise here. In our [arXiv paper](https://arxiv.org/pdf/2502.05311) (Section 4.4.2 Preprocessing Incoming Data), we describe this set of operations as a **`preprocessing`** step. This `preprocessing` aims to bring the data into a "normal" or consistent and optimized state for storage and querying. The specific `preprocessing` steps involved are:

- **Handling Empty Structs:** ParquetDB ensures that empty dictionaries (structs) are handled appropriately with the use of a dummy variable to prevent errors when writing to Parquet files, as some engines may raise errors in such cases.
- **Flattening Nested Structures:** To enhance ease of processing and querying, ParquetDB can flatten nested data structures (e.g., structs within structs, dictionaries with nested fields). This process creates a flat tabular representation where new columns are generated for each nested field. These new columns follow a clear naming convention, such as `parent.child1.child2`, reflecting the original hierarchy (e.g., `address.city`, `address.postal_code`). This simplification facilitates data access, analysis, schema change detection, and compatibility with various query engines. ParquetDB also retains the capability to reconstruct the nested structure if it's more suitable for specific analytical tasks.
- **Column Reordering:** ParquetDB alphabetically sorts columns. This consistent ordering simplifies the detection of schema changes and updates, as new columns are automatically placed in their correct alphabetical position, making schema differences easier to identify.
- **Schema Alignment:** When new data is ingested, ParquetDB compares its schema against the existing dataset's schema. Depending on the operation and any detected differences, ParquetDB will either update the overall dataset schema or cast the incoming data to conform to the existing schema, ensuring consistency.

> Could you please define what you mean by a "normal state" of data?

Therefore, the "normal state" of data in ParquetDB refers to data that has undergone these `preprocessing` steps, resulting in a clean, consistent, and optimized format for efficient storage and analysis. We agree that "preprocessing" is likely a more accurate and less ambiguous term for these operations.

### On the Statement of Need

> Can you define the use case precisely? What are the characteristics of the data and of the workload that make the package useful in particular?

The primary value of ParquetDB lies in the combination of the `preprocessing` steps detailed above and the robust management of the data's state over time. Essentially, ParquetDB acts as a wrapper around `pyarrow`, incorporating crucial preprocessing functionalities. These steps are designed to save users from writing significant boilerplate code and from navigating the complexities of managing and updating immutable `pyarrow` tables, especially when dealing with evolving schemas and nested data structures.

For instance, if a user were to rely solely on raw `pyarrow` or `pandas` for similar tasks, they would be responsible for:
- Manually handling schema alignment between incoming data and existing datasets.
- Implementing their own row indexing mechanisms.
- Managing data updates, which becomes particularly challenging with nested structures and the immutability of Parquet files (requiring rewriting files or partitions).

When developing ParquetDB, we evaluated several existing tools, including [Pandas](https://pandas.pydata.org/), [Dask](https://www.dask.org/), [DuckDB](https://duckdb.org/), [PySpark](https://spark.apache.org/docs/latest/api/python/index.html), [Delta-rs](https://github.com/delta-io/delta-rs), and [Polars](https://pola.rs/). Our primary use case at the time involved complex materials science data, which is often characterized by deeply nested structures and evolving schemas as research progresses. Our key requirements were:

- **Schema Evolvability:** The ability to adapt the schema over time without defining a rigid structure upfront is crucial for iterative research.
- **Handling Nested Structures:** Robust support for complex, nested data was a primary driver.
- **Performance:** The solution needed to be efficient for data storage and retrieval.
- **"Classically Serverless":** No dependency on a network connection to a database server. This was particularly important for our work on High-Performance Computing (HPC) clusters where compute nodes often lack internet connectivity. Setting up local servers on other compute nodes felt clunky and inefficient.
- **Ability to Handle Table and Field Metadata:** Storing and managing metadata associated with tables and individual fields was also a requirement.

While `MongoDB` might seem like a candidate given its flexibility with schemas and nested data, it did not meet our "serverless" requirement and performance considerations for our specific analytical workloads. This led us to explore file-based solutions and build ParquetDB.

Here's a brief comparison highlighting why other tools didn't fully meet our specific needs, leading to the development of ParquetDB:

* **`Pandas`:** While excellent for in-memory data manipulation and capable of reading/writing Parquet files, `pandas` itself doesn't inherently manage schema evolution across multiple Parquet files or simplify the complexities of updating nested data within a Parquet-based dataset. Users would need to build these management layers themselves.
- **`Dask`:** Similar to `pandas`, `Dask` excels at parallel processing of large datasets but doesn't provide the specific schema evolution and nested data update management layer that ParquetDB offers. The onus of managing these aspects remains with the user.
- **`DuckDB`:** DuckDB is an impressive embedded analytical database and a strong contender for many use cases. However, at the time of our initial development and based on our exploration, its support for schema evolution, particularly with deeply nested structures, and atomic updates of individual fields within nested structures did not align with our requirements. For example, updating a single child field within a nested struct often required rewriting the entire struct.
- **`Polars`:** A very fast DataFrame library, but like `pandas` and `Dask`, it focuses on data processing rather than the state management and schema evolution for datasets persisted as a collection of Parquet files with complex nested data. Our concerns regarding schema evolution and atomic updates of nested structures were similar to those with DuckDB.
- **`Delta-rs (Delta Lake)`:** Delta Lake offers ACID transactions, schema enforcement, and scalable data handling, which are very valuable. However, our primary challenge was the nuanced handling of schema *evolution* for *deeply nested* data and fine-grained updates within those nested structures, which we found ParquetDB could be tailored to address more directly for our use case.

ParquetDB was designed to fill this gap by providing a higher-level API that abstracts these complexities. The core contribution is the streamlined handling of these more intricate data management scenarios, making it easier for users to work with evolving, nested Parquet datasets. While the backend could theoretically leverage components from some of these other tools, the unique value of ParquetDB lies in its opinionated approach to preprocessing and state management that simplifies the user experience for these specific challenges.

> Could you please write something short about why I should not use the well-known Pandas and Dask instead?

If your data is characterized by simple, flat structures, or perhaps only shallowly nested data with a stable schema, then tools like `DuckDB` or managing Parquet files directly with `pandas` or `pyarrow` might indeed be sufficient and more straightforward. However, ParquetDB becomes particularly useful when dealing with:

- Data with complex and potentially deep nesting.
- Schemas that are expected to evolve over time (e.g., adding new fields, modifying nested structures) in an iterative research or development context.
- The need for a "serverless" solution that manages collections of Parquet files as a coherent, updatable dataset.

We believe ParquetDB offers a valuable solution for users facing these specific data characteristics and workloads by simplifying the intricate aspects of managing evolving, nested data in a Parquet-based environment.

Thank you again for your review and insightful questions. We hope this clarifies our intent and the specific needs ParquetDB aims to address. We look forward to your further feedback.

Best Regards,

Logan Lang
