### 2025-05-07 3633767
test: Refactor tests for ParquetDB to improve data validation
Updated test cases in test_parquetdb.py to enhance data validation
by asserting the shape and content of the returned arrays. This
change ensures that the tests accurately reflect the expected
behavior of the database operations. Previous tests did not take into account that multithreaded write has a chance to preserve row order. The ultimate source of truth is the "id" that indexes all the rows
