+++ tests/test_parquetdb.py
@@ -653,78 +653,97 @@ class TestParquetDB(unittest.TestCase):
         data_1 = [{"pbc": [1, 0, 0]}, {"pbc": [0, 1, 0]}]
         self.db.create(data_1)
 
-        table = self.db.read()
+        table = self.db.read(ids=[0])
+        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [[1, 0, 0]]
+        table = self.db.read(ids=[1])
+        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [[0, 1, 0]]
 
-        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [
-            [1, 0, 0],
-            [0, 1, 0],
-        ]
         data_2 = [{"id": 0, "density": 5}]
 
         self.db.update(data_2)
         table = self.db.read(ids=[1])
         df = table.to_pandas()
+
         logger.debug(f"Dataframe: \n {df}")
-        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [
-            [0, 1, 0],
-        ]
+        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [[0, 1, 0]]
 
         table = self.db.read(ids=[0])
-
-        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [
-            [1, 0, 0],
-        ]
+        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [[1, 0, 0]]
 
     def test_update_maintains_existing_extension_arrays_batches(self):
         data_1 = [{"pbc": [1, 0, 0]}, {"pbc": [0, 1, 0]}]
         self.db.create(data_1)
 
-        table = self.db.read()
+        table = self.db.read(ids=[0])
+        pbc_array = table["pbc"].combine_chunks().to_numpy_ndarray()
+        assert pbc_array.tolist() == [[1, 0, 0]]
+        assert pbc_array.shape == (1, 3)
 
-        assert table["pbc"].combine_chunks().to_numpy_ndarray().tolist() == [
-            [1, 0, 0],
-            [0, 1, 0],
-        ]
+        table = self.db.read(ids=[1])
+        pbc_array = table["pbc"].combine_chunks().to_numpy_ndarray()
+        assert pbc_array.tolist() == [[0, 1, 0]]
+        assert pbc_array.shape == (1, 3)
+
+        table = self.db.read()
+        pbc_array = table["pbc"].combine_chunks().to_numpy_ndarray()
+
+        assert pbc_array.shape == (2, 3)
+
         data_2 = [{"id": 0, "density": 5}]
 
         self.db.update(
             data_2,
             normalize_config=NormalizeConfig(load_format="batches", batch_size=1),
         )
+
+        table = self.db.read(ids=[0])
+        pbc_array = table["pbc"].combine_chunks().to_numpy_ndarray()
+        assert pbc_array.tolist() == [[1, 0, 0]]
+        assert table["density"].combine_chunks().to_pylist() == [5]
+        assert pbc_array.shape == (1, 3)
+
+        table = self.db.read(ids=[1])
+        pbc_array = table["pbc"].combine_chunks().to_numpy_ndarray()
+
+        assert pbc_array.tolist() == [[0, 1, 0]]
+        assert table["density"].combine_chunks().to_pylist() == [None]
+        assert pbc_array.shape == (1, 3)
+
         table = self.db.read()
-        df = table.to_pandas()
-        logger.debug(f"Dataframe: \n {df}")
-
-        array = table["pbc"].combine_chunks().to_numpy_ndarray().tolist()
-
[…]
