### 2025-05-07 035b070
refactor(tests): Update assertions in test_parquetdb for clarity
Modified assertions in test_parquetdb.py to use parentheses for
logical conditions, improving readability and correctness. This
change ensures that the tests accurately validate the expected
output from the DataFrame, enhancing the reliability of the test
suite.
