### 2025-05-07 1bb1e28
test: Update assertions in test_parquetdb for type validation
Modified assertions in test_parquetdb.py to check the types of
returned values instead of their specific values. This change
ensures that the tests validate the expected data types,
improving the robustness of the test suite and ensuring
correctness in the handling of file sizes and serialized
metadata size.
