+++ tests/test_parquetdb.py
@@ -1361,12 +1361,9 @@ class TestParquetDB(unittest.TestCase):
 
     def test_get_file_sizes(self):
         self.db.create(data=self.test_data)
-        logger.debug(self.db.get_file_sizes())
-        file_sizes = {"test_db_0.parquet": 0.0012388229370117188}
-        assert np.isclose(
-            self.db.get_file_sizes()["test_db_0.parquet"],
-            file_sizes["test_db_0.parquet"],
-        )
+        logger.debug(f"file_sizes: {self.db.get_file_sizes()}")
+        assert isinstance(self.db.get_file_sizes(), dict)
+        assert isinstance(self.db.get_file_sizes()["test_db_0.parquet"], float)
 
     def test_get_parquet_column_metadata_per_file(self):
         self.db.create(data=self.test_data)
@@ -1408,7 +1405,7 @@ class TestParquetDB(unittest.TestCase):
         logger.debug(
             f"serialized_metadata_size_per_file: {self.db.serialized_metadata_size_per_file}"
         )
-        assert self.db.serialized_metadata_size_per_file[0] == 717
+        assert isinstance(self.db.serialized_metadata_size_per_file[0], int)
 
     def test_summary(self):
         self.db.create(data=self.test_data)
