### 2025-05-07 0a9dfa1
docs: Update example titles in index.rst for clarity
Renamed example titles in index.rst to better reflect their content.
Changed "Creating a GraphDB" to "Managing Graphs in ParquetGraphDB"
and added "Graph Generators in ParquetGraphDB" to enhance user
understanding of the examples provided.
