### 2025-05-09 #7932 @perdelt
Dear @lllangWV

Thank you very much for the detailed explanations!

I understand. Yes, that makes perfect sense.

I suggest condensing the requirements section you wrote above (*Our key requirements were*) into one or two sentences and placing it in the first "Summary" section of the paper. This way, anyone with similar requirements can immediately see "this package is for me."

I also suggest briefly including the comparisons to other tools like Pandas, Dask, etc. in the "Statement of Need" section. It would again help to recognize that ParquetDB has features that other tools don't have and that the package primarily addresses the people who need these features. A more detailed discussion might be something for the docs.
