### 2025-05-10 8374cb4
docs: Add API documentation for graph components
Introduced new documentation files for EdgeGenerator, EdgeStore,
GeneratorStore, GraphDB, NodeGenerator, and NodeStore. Each file
provides a detailed description of the respective class or function,
including its purpose and usage within the MatGraphDB framework.
This addition enhances the overall documentation and aids users
in understanding the core API functionalities.
