+++ docs/source/03_api/graph/generator_store.rst
@@ -1,0 +1,9 @@
+GeneratorStore
+========================
+
+- :class:`GeneratorStore <matgraphdb.core.generator_store.GeneratorStore>` - A store for managing generator functions in a graph database. This class handles serialization, storage, and loading of functions that generate edges between nodes.
+
+.. autosummary::
+   :toctree: _autosummary
+
+   matgraphdb.core.generator_store.GeneratorStore
