+++ docs/source/03_api/graph/_autosummary/matgraphdb.core.graph_db.GraphDB.rst
@@ -1,0 +1,70 @@
+﻿matgraphdb.core.graph\_db.GraphDB
+=================================
+
+.. currentmodule:: matgraphdb.core.graph_db
+
+.. autoclass:: GraphDB
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~GraphDB.__init__
+      ~GraphDB.add_edge_generator
+      ~GraphDB.add_edge_store
+      ~GraphDB.add_edge_type
+      ~GraphDB.add_edges
+      ~GraphDB.add_generator_dependency
+      ~GraphDB.add_node_generator
+      ~GraphDB.add_node_store
+      ~GraphDB.add_node_type
+      ~GraphDB.add_nodes
+      ~GraphDB.construct_table
+      ~GraphDB.delete_edges
+      ~GraphDB.delete_nodes
+      ~GraphDB.edge_exists
+      ~GraphDB.edge_is_empty
+      ~GraphDB.generator_consistency_check
+      ~GraphDB.get_edge_store
+      ~GraphDB.get_generator_dependency_graph
+      ~GraphDB.get_generator_type
+      ~GraphDB.get_node_store
+      ~GraphDB.get_nodes
+      ~GraphDB.list_edge_types
+      ~GraphDB.list_node_types
+      ~GraphDB.node_exists
+      ~GraphDB.node_is_empty
+      ~GraphDB.normalize_all_edges
+      ~GraphDB.normalize_all_nodes
+      ~GraphDB.normalize_edges
+      ~GraphDB.normalize_nodes
+      ~GraphDB.read_edges
+      ~GraphDB.read_nodes
+      ~GraphDB.remove_edge_store
+      ~GraphDB.remove_edge_type
+      ~GraphDB.remove_node_store
+      ~GraphDB.remove_node_type
+      ~GraphDB.run_edge_generator
+      ~GraphDB.run_node_generator
+      ~GraphDB.summary
+      ~GraphDB.update_edges
+      ~GraphDB.update_nodes
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~GraphDB.n_edge_types
+      ~GraphDB.n_edges_per_type
+      ~GraphDB.n_node_types
+      ~GraphDB.n_nodes_per_type
+   
+   
