+++ docs/source/03_api/graph/_autosummary/matgraphdb.core.nodes.node_generator.rst
@@ -1,0 +1,6 @@
+﻿matgraphdb.core.nodes.node\_generator
+=====================================
+
+.. currentmodule:: matgraphdb.core.nodes
+
+.. autofunction:: node_generator
