+++ docs/source/03_api/graph/node_store.rst
@@ -1,0 +1,9 @@
+NodeStore
+========================
+
+- :class:`NodeStore <matgraphdb.core.nodes.NodeStore>` - A store for managing node features in a graph database. This class handles data storage, retrieval, querying, schema evolution, and complex data type management through an intuitive API that wraps PyArrow's functionality.
+
+.. autosummary::
+   :toctree: _autosummary
+
+   matgraphdb.core.nodes.NodeStore
