+++ docs/source/03_api/graph/edge_generator.rst
@@ -1,0 +1,9 @@
+EdgeGenerator
+========================
+
+- :func:`edge_generator <matgraphdb.core.edges.edge_generator>` - A decorator that validates the input arguments of a function and converts them into a dataframe.
+
+.. autosummary::
+   :toctree: _autosummary
+
+   matgraphdb.core.edges.edge_generator
