+++ docs/source/03_api/graph/node_generator.rst
@@ -1,0 +1,9 @@
+NodeGenerator
+========================
+
+- :func:`node_generator <matgraphdb.core.nodes.node_generator>` - A decorator that validates the input arguments of a function and converts them into a dataframe.
+
+.. autosummary::
+   :toctree: _autosummary
+
+   matgraphdb.core.nodes.node_generator
