### 2025-05-10 a0ebd94
docs: Update API documentation to reflect parquetdb namespace
Replaced instances of 'matgraphdb' with 'parquetdb' in API documentation
files for EdgeGenerator, EdgeStore, GeneratorStore, GraphDB, NodeGenerator,
and NodeStore. This change aligns the documentation with the updated
module structure and ensures consistency across the API references.
