### 2025-05-10 dbd4d62
docs: Update API documentation to reflect new parquetdb structure
Replaced instances of 'parquetdb.core' with 'parquetdb.graph' in the API
documentation files for EdgeGenerator, EdgeStore, GeneratorStore, GraphDB,
NodeGenerator, and NodeStore. This change aligns the documentation with
the updated module structure and ensures consistency across the API
references.
