+++ docs/source/03_api/graph/edge_store.rst
@@ -1,9 +1,9 @@
 Edge Store
 ========================
 
-- :class:`EdgeStore <parquetdb.core.edges.EdgeStore>` - The main interface class that provides database-like operations over Parquet files. This class handles data storage, retrieval, querying, schema evolution, and complex data type management through an intuitive API that wraps PyArrow's functionality.
+- :class:`EdgeStore <parquetdb.graph.edges.EdgeStore>` - The main interface class that provides database-like operations over Parquet files. This class handles data storage, retrieval, querying, schema evolution, and complex data type management through an intuitive API that wraps PyArrow's functionality.
 
 .. autosummary::
    :toctree: _autosummary
 
-   parquetdb.core.edges.EdgeStore
+   parquetdb.graph.edges.EdgeStore
