+++ docs/source/03_api/graph/node_generator.rst
@@ -1,9 +1,9 @@
 NodeGenerator
 ========================
 
-- :func:`node_generator <parquetdb.core.nodes.node_generator>` - A decorator that validates the input arguments of a function and converts them into a dataframe.
+- :func:`node_generator <parquetdb.graph.nodes.node_generator>` - A decorator that validates the input arguments of a function and converts them into a dataframe.
 
 .. autosummary::
    :toctree: _autosummary
 
-   parquetdb.core.nodes.node_generator
+   parquetdb.graph.nodes.node_generator
