+++ docs/source/03_api/graph/edge_generator.rst
@@ -1,9 +1,9 @@
 EdgeGenerator
 ========================
 
-- :func:`edge_generator <parquetdb.core.edges.edge_generator>` - A decorator that validates the input arguments of a function and converts them into a dataframe.
+- :func:`edge_generator <parquetdb.graph.edges.edge_generator>` - A decorator that validates the input arguments of a function and converts them into a dataframe.
 
 .. autosummary::
    :toctree: _autosummary
 
-   parquetdb.core.edges.edge_generator
+   parquetdb.graph.edges.edge_generator
