+++ docs/source/03_api/graph/graphdb.rst
@@ -1,9 +1,9 @@
 GraphDB
 ========================
 
-- :class:`GraphDB <parquetdb.core.graph_db.GraphDB>` - A manager for a graph storing multiple node types and edge types. Each node type and edge type is backed by a separate ParquetDB instance (wrapped by NodeStore or EdgeStore).
+- :class:`GraphDB <parquetdb.graph.graph_db.GraphDB>` - A manager for a graph storing multiple node types and edge types. Each node type and edge type is backed by a separate ParquetDB instance (wrapped by NodeStore or EdgeStore).
 
 .. autosummary::
    :toctree: _autosummary
 
-   parquetdb.core.graph_db.GraphDB
+   parquetdb.graph.graph_db.GraphDB
