+++ docs/source/03_api/graph/node_store.rst
@@ -1,9 +1,9 @@
 NodeStore
 ========================
 
-- :class:`NodeStore <parquetdb.core.nodes.NodeStore>` - A store for managing node features in a graph database. This class handles data storage, retrieval, querying, schema evolution, and complex data type management through an intuitive API that wraps PyArrow's functionality.
+- :class:`NodeStore <parquetdb.graph.nodes.NodeStore>` - A store for managing node features in a graph database. This class handles data storage, retrieval, querying, schema evolution, and complex data type management through an intuitive API that wraps PyArrow's functionality.
 
 .. autosummary::
    :toctree: _autosummary
 
-   parquetdb.core.nodes.NodeStore
+   parquetdb.graph.nodes.NodeStore
