### 2025-05-10 089137f
docs: Update GraphDB references to ParquetGraphDB in documentation
Replaced instances of 'GraphDB' with 'ParquetGraphDB' in the API
documentation to accurately reflect the updated class name.
Additionally, created a new documentation file for ParquetGraphDB
to provide detailed information about its functionality and usage.
