+++ docs/source/03_api/graph/index.rst
@@ -16,7 +16,7 @@ The core components include:
 
 - :func:`node_generator <parquetdb.graph.nodes.node_generator>` - A decorator that validates the input arguments of a function and converts them into a dataframe.
 
-- :class:`GraphDB <parquetdb.graph.graph_db.GraphDB>` - A manager for a graph storing multiple node types and edge types. Each node type and edge type is backed by a separate ParquetDB instance (wrapped by NodeStore or EdgeStore).
+- :class:`ParquetGraphDB <parquetdb.graph.parquet_graphdb.ParquetGraphDB>` - A manager for a graph storing multiple node types and edge types. Each node type and edge type is backed by a separate ParquetDB instance (wrapped by NodeStore or EdgeStore).
 
 .. toctree::
    :maxdepth: 2
@@ -26,4 +26,4 @@ The core components include:
    edge_store
    edge_generator
    generator_store
-   graphdb
+   parquet_graphdb
