### 2025-05-10 deb2767
docs: Rename GraphDB to ParquetGraphDB in documentation
Updated references in the documentation to reflect the class name change
from 'GraphDB' to 'ParquetGraphDB'. This ensures consistency and accuracy
in the API documentation for users.
