+++ docs/source/03_api/graph/parquet_graphdb.rst
@@ -1,9 +1,9 @@
-GraphDB
+ParquetGraphDB
 ========================
 
-- :class:`GraphDB <parquetdb.graph.graph_db.GraphDB>` - A manager for a graph storing multiple node types and edge types. Each node type and edge type is backed by a separate ParquetDB instance (wrapped by NodeStore or EdgeStore).
+- :class:`ParquetGraphDB <parquetdb.graph.parquet_graphdb.ParquetGraphDB>` - A manager for a graph storing multiple node types and edge types. Each node type and edge type is backed by a separate ParquetDB instance (wrapped by NodeStore or EdgeStore).
 
 .. autosummary::
    :toctree: _autosummary
 
-   parquetdb.graph.graph_db.GraphDB
+   parquetdb.graph.parquet_graphdb.ParquetGraphDB
