### 2025-05-11 896b35e
refactor: Remove code depedency to Stores and refactor paths to use pathlib.Path
Previousaly, NodeStore, GeneratoreStore, and EdgeStore all stored their class name and module path in the metadata. This is bad because if the cod eis updated all subsequent data will be outdated. Instead, Users can provide the specify custom types on Initating Stores/ ParquetGraphDB.

Modified EdgeStore and NodeStore constructors to accept storage_path as
either a string or Path object. This change improves flexibility in
handling file paths. Updated related methods to ensure compatibility
with the new storage_pa
[…]
