+++ parquetdb/graph/parquet_graphdb.py
@@ -1,23 +1,17 @@
-import importlib
 import json
 import logging
 import os
 import shutil
-import time
-from glob import glob
 from pathlib import Path
-from typing import Callable, Dict, List, Union
+from typing import Callable, Dict, List, Type, Union
 
-import pandas as pd
 import pyarrow as pa
 import pyarrow.compute as pc
-from pyarrow import parquet as pq
 
 from parquetdb import ParquetDB
 from parquetdb.graph.edges import EdgeStore
 from parquetdb.graph.generator_store import GeneratorStore
 from parquetdb.graph.nodes import NodeStore
-from parquetdb.utils import pyarrow_utils
 from parquetdb.utils.log_utils import set_verbose_level
 
 logger = logging.getLogger(__name__)
@@ -31,7 +25,11 @@ class ParquetGraphDB:
     """
 
     def __init__(
-        self, storage_path: str, load_custom_stores: bool = True, verbose: int = 1
+        self,
+        storage_path: Union[str, Path],
+        node_store_types: Dict[str, Type[NodeStore]] = None,
+        edge_store_types: Dict[str, Type[EdgeStore]] = None,
+        verbose: int = 1,
     ):
         """
         Parameters
@@ -44,32 +42,53 @@ class ParquetGraphDB:
         set_verbose_level(verbose)
 
         logger.info(f"Initializing GraphDB at root path: {storage_path}")
-        self.storage_path = storage_path
-
-        self.nodes_path = os.path.join(self.storage_path, "nodes")
-        self.edges_path = os.path.join(self.storage_path, "edges")
-        self.edge_generators_path = os.path.join(self.storage_path, "edge_generators")
-        self.node_generators_path = os.path.join(self.storage_path, "node_generators")
-        self.graph_path = os.path.join(self.storage_path, "graph")
-        self.generator_dependency_json = os.path.join(
-            self.storage_path, "generator_dependency.json"
+        self.storage_path = (
+            Path(storage_path) if isinstance(storage_path, str) else storage_path
         )
 
-        self.graph_name = os.path.basename(self.storage_path)
+        self.nodes_path = self.storage_path / "nodes"
+        self.edges_path = self.storage_path / "edges"
+        self.edge_generators_path = self.storage_path / "edge_generators"
+        self.node_generators_path = self.storage_path / "node_generators"
+        self.graph_path = self.storage_path / "graph"
+        self.generator_dependency_json = self.storage_path / "generator_dependency.json"
+
+        self.graph_name = self.storage_path.name
 
         # Create directories if they don't exist
-        os.makedirs(self.nodes_path, exist_ok=True)
-        os.makedirs(self.edges_path, exist_ok=True)
-        os.makedirs(self.edge_generators_path, exist_ok=True)
-        os.makedirs(self.graph_path, exist_ok=True)
+        self.nodes_path.mkdir(parents=True, exist_ok=True)
+        self.edges_path.mkdir(parents=True, exist_ok=True)
+        self.edge_generators_path.mkdir(parents=True, exist_ok=True)
+        self.graph_path.mkdir(parents=True, exist_ok=True)
 
         logger.debug(f"Node directory: {self.nodes_path}")
         logger.debug(f"Edge directory: {self.edges_path}")
         logger.debug(f"Graph directory: {self.graph_path}")
 
         #  Initialize empty dictionaries for stores, load existing stores
-        self.node_stores = self._load_existing_node_stores(load_custom_stores)
-        self.edge_stores = self._load_existing_edge_stores(load_custom_stores)
+        self.node_stores = {}
+        self.edge_stores = {}
+
+        node_store_types = {} if node_store_types is None else node_store_types
[…]
