### 2025-05-11 6c498cb
refactor: Simplify initialization of setup_kwargs and initialize_kwargs
Updated EdgeStore and NodeStore constructors to streamline the
initialization of setup_kwargs and initialize_kwargs. This change
ensures that these variables are assigned an empty dictionary only
if they are None, improving code readability and maintainability.
