### 2025-05-12 d156c7d
refactor(tests): Remove unused metadata assertions in edge store tests
Eliminated assertions for metadata in the edge store initialization test
to streamline the test code. This change reduces redundancy and focuses
on the essential aspects of the edge store's functionality.

No functional changes were made to the edge store itself.
