### 2025-05-12 06b6cfb
refactor(parquetdb): Update ParquetDB to use pathlib.Path for file handling and added type hints
Refactored the ParquetDB class to replace string-based file path handling
with pathlib.Path objects. This change improves path manipulation and
ensures better compatibility with modern Python practices.

Updated method signatures and internal logic to accommodate the new
Path type

Additionally, added type hints for better clarity and consistency across
the codebase.

No functional changes were made to the ParquetDB's core functionality.
