+++ parquetdb/core/parquetdb.py
@@ -1,13 +1,13 @@
-import copy
 import itertools
 import logging
 import os
 import shutil
-import time
 import types
+from collections.abc import Iterable
 from dataclasses import dataclass
 from glob import glob
-from typing import Callable, Dict, List, Optional, Union
+from pathlib import Path
+from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 
 import dill
 import pandas as pd
@@ -160,7 +160,7 @@ class LoadConfig:
 class ParquetDB:
     def __init__(
         self,
-        db_path,
+        db_path: Union[str, Path],
         initial_fields: List[pa.Field] = None,
         serialize_python_objects: bool = False,
         use_multiprocessing: bool = False,
@@ -197,6 +197,7 @@ class ParquetDB:
         >>> fields = [pa.field('name', pa.string()), pa.field('age', pa.int32())]
         >>> db = ParquetDB(db_path='/path/to/db', initial_fields=fields)
         """
+
         logger.info(f"Initializing ParquetDB with db_path: {db_path}")
 
         logger.info(f"verbose: {verbose}")
@@ -205,8 +206,8 @@ class ParquetDB:
         self.normalize_config = normalize_config
         self.load_config = load_config
 
-        self._db_path = db_path
-        os.makedirs(self._db_path, exist_ok=True)
+        self._db_path = Path(db_path)
+        self._db_path.mkdir(parents=True, exist_ok=True)
         self._serialize_python_objects = serialize_python_objects
 
         if initial_fields is None:
@@ -233,19 +234,19 @@ class ParquetDB:
         return self.summary(show_column_names=False)
 
     @property
-    def db_path(self):
+    def db_path(self) -> Path:
         """
         Get the database path.
 
         Returns
         -------
-        str
+        Path
             The path to the database directory.
         """
         return self._db_path
 
     @property
-    def dataset_name(self):
+    def dataset_name(self) -> str:
         """
         Get the dataset name.
 
@@ -254,10 +255,10 @@ class ParquetDB:
         str
             The name of the dataset, derived from the database path.
         """
-        return os.path.basename(self._db_path)
+        return self._db_path.name
 
     @property
-    def basename_template(self):
+    def basename_template(self) -> str:
         """
         Get the template for parquet file basenames.
 
@@ -269,7 +270,7 @@ class ParquetDB:
         return f"{self.dataset_name}_{{i}}.parquet"
 
     @property
-    def n_columns(self):
+    def n_columns(self) -> int:
         """
         Get the number of columns in the database.
 
@@ -281,19 +282,19 @@ class ParquetDB:
         return len(self.columns)
 
     @property
-    def columns(self):
+    def columns(self) -> List[str]:
         """
         Get the column names in the database.
 
         Returns
         -------
-        list
+        List[str]
             List of column names.
         """
         return self.get_field_names()
 
     @property
-    def n_rows(self):
+    def n_rows(self) -> int:
         """
         Get the total number of rows in the database.
 
@@ -306,7 +307,7 @@ class ParquetDB:
         return ds.count_rows()
 
     @property
-    def n_files(self):
+    def n_files(self) -> int:
         """
         Get the number of parquet files in the database.
 
@@ -318,50 +319,54 @@ class ParquetDB:
         return len(self.get_current_files())
 
     @property
-    def n_rows_per_file(self):
+    def n_rows_per_file(self) -> Dict[str, int]:
         """
         Get the number of rows in each parquet file.
 
         Returns
         -------
-        dict
-            Dictionary mapping filenames to their row counts.
+        Dict[str, int]
+            Returns a dictionary detailing the number of rows in each processed Parquet file.
+            Structure: {filename: n_rows}
         """
[…]
