### 2025-05-12 2b4b28b
bug: Fixed bug in rename_dataset
