### 2025-05-12 e11cd49
feat(node): Simplified source of truth for node_type. Added feature to easily change node type
Additionally, implemented the `set_node_type` method to change the
node type and update the associated metadata accordingly.

Updated tests to verify the functionality of the new method and property.
