### 2025-05-13 67cf7df
feat(docs): Update paper.md to clarify ParquetDB's design motivations
Expanded the summary and statement of need section to emphasize the need for schema evolvability
and support for complex nested data structures in iterative research workflows.
This change aims to better articulate the unique advantages of ParquetDB
over traditional databases and other data manipulation libraries.
