+++ joss/ParquetDB.bib
@@ -12,6 +12,14 @@
   keywords = {database,database management,relational database,SQL,SQLite}
 }
 
+@misc{DaskScalePythona,
+  title = {Dask {\textbar} {{Scale}} the {{Python}} Tools You Love},
+  urldate = {2025-05-13},
+  abstract = {Dask is a flexible open-source Python library for parallel computing maintained by OSS contributors across dozens of companies including Anaconda, Coiled, SaturnCloud, and nvidia.},
+  howpublished = {https://www.dask.org/},
+  langid = {english}
+}
+
 @article{guoMongoDBsJavaScriptFuzzer2017,
   title = {{{MongoDB}}'s {{JavaScript}} Fuzzer},
   author = {Guo, Robert},
@@ -77,6 +85,12 @@
   abstract = {Accelerating the discovery of advanced materials is essential for human welfare and sustainable, clean energy. In this paper, we introduce the Materials Project (www.materialsproject.org), a core program of the Materials Genome Initiative that uses high-throughput computing to uncover the properties of all known inorganic materials. This open dataset can be accessed through multiple channels for both interactive exploration and data mining. The Materials Project also seeks to create open-source platforms for developing robust, sophisticated materials analyses. Future efforts will enable users to perform ``rapid-prototyping'' of new materials in silico, and provide researchers with new avenues for cost-effective, data-driven materials design.}
 }
 
+@misc{PandasPythonData,
+  title = {Pandas - {{Python Data Analysis Library}}},
+  urldate = {2025-05-13},
+  howpublished = {https://pandas.pydata.org/}
+}
+
 @book{pascalPracticalIssuesDatabase2000,
   title = {Practical {{Issues}} in {{Database Management}}: {{A Refernce}} for the {{Thinking Practitioner}}},
   shorttitle = {Practical {{Issues}} in {{Database Management}}},
@@ -105,44 +119,20 @@
   langid = {english}
 }
 
-
-@misc{Parquet,
-  title = {Parquet},
-  journal = {Apache Parquet},
-  urldate = {2024-10-21},
-  abstract = {The Apache Parquet Website},
-  howpublished = {https://parquet.apache.org/},
+@misc{Polars,
+  title = {Polars},
+  urldate = {2025-05-13},
+  abstract = {DataFrames for the new era},
+  howpublished = {https://www.pola.rs/},
   langid = {english}
 }
 
-
-@misc{WelcomeApacheSoftware,
-  title = {Welcome to {{The Apache Software Foundation}}},
-  urldate = {2025-04-28},
-  howpublished = {https://www.apache.org/}
-}
-
-
-@misc{HDF5PythonH5py,
-  title = {{{HDF5}} for {{Python}} --- H5py 3.13.0 Documentation},
-  urldate = {2025-04-28},
-  howpublished = {https://docs.h5py.org/en/stable/index.html}
-}
-
-@misc{langParquetDBLightweightPython2025,
-  title = {{{ParquetDB}}: {{A Lightweight Python Parquet-Based Database}}},
-  shorttitle = {{{ParquetDB}}},
-  author = {Lang, Logan and Hernandez, Eduardo and Choudhary, Kamal and Romero, Aldo H.},
-  year = {2025},
-  month = feb,
-  number = {arXiv:2502.05311},
-  eprint = {2502.05311},
-  primaryclass = {cs},
-  publisher = {arXiv},
-  doi = {10.48550/arXiv.2502.05311},
-  urldate = {2025-02-14},
[…]
