### 2025-05-13 60e77e2
feat(docs): Add references for Apache Parquet and Apache Software Foundation
Included new entries in the bibliography for Apache Parquet and the
Apache Software Foundation. This addition aims to provide readers with
relevant resources and enhance the documentation's comprehensiveness.
