+++ joss/ParquetDB.bib
@@ -91,6 +91,15 @@
   howpublished = {https://pandas.pydata.org/}
 }
 
+@misc{Parquet,
+  title = {Parquet},
+  journal = {Apache Parquet},
+  urldate = {2024-10-21},
+  abstract = {The Apache Parquet Website},
+  howpublished = {https://parquet.apache.org/},
+  langid = {english}
+}
+
 @book{pascalPracticalIssuesDatabase2000,
   title = {Practical {{Issues}} in {{Database Management}}: {{A Refernce}} for the {{Thinking Practitioner}}},
   shorttitle = {Practical {{Issues}} in {{Database Management}}},
@@ -135,4 +144,10 @@
   abstract = {DuckDB is an in-process SQL OLAP database management system. Simple, feature-rich, fast \& open source.},
   howpublished = {https://duckdb.org/},
   langid = {english}
+}
+
+@misc{WelcomeApacheSoftware,
+  title = {Welcome to {{The Apache Software Foundation}}},
+  urldate = {2025-04-28},
+  howpublished = {https://www.apache.org/}
 }
