### 2025-05-13 #7932 @lllangWV
Dear @perdelt

I've addressed your suggestions as follows:

- **Summary Section Update**: As you recommended, I've condensed the key requirements that motivated ParquetDB into two sentences and incorporated them into the "Summary" section of the paper. This should help readers with similar needs quickly identify the relevance of the package.

> A key driver for its development was the need to support iterative research workflows, requiring schema evolvability, the ability to manage complex and evolving nested data structures without predefined rigidity, and the ability to handle table and field level metadata. Additionally, its "classically serverless" nature was a crucial design point for deployment in environments such as HPC clusters with limited connectivity.

- **Statement of Need Update**: I've also added a brief comparison to other tools like Pandas, Dask, DuckDB, and Polars within the "Statement of Need" section. This clarifies ParquetDB's specific features and the niche it aims to fill.

> While powerful dataframe manipulation libraries like Pandas[@PandasPythonData], Dask[@DaskScalePythona], and Polars[@Polars], or embedded analytical databases such as DuckDB[@userInprocessSQLOLAP], are invaluable for many tasks, they may not holistically address the specific needs that motivated ParquetDB. For researchers dealing with evolving, complexly nested scientific data, ParquetDB offers a more streamlined approach to schema evolvability and native Python object persistence directly within a serverless Parquet-based ecosystem. This focus distinguishes it from tools that might require more manual setup for schema management across multiple files, or lack the same emphasis on integrated metadata handling and a 'classically serverless' model for environments like HPC clusters.

- **Documentation**: For a more detailed discussion of the motivation and comparison with other tools, I have created a new "ParquetDB Motivation" section in the documentation, as you suggested.

> https://parquetdb.readthedocs.io/en/latest/02_internal/00%20-%20parquetdb_motivation.html

I believe these revisions have strengthened the paper and better highlight the specific problems ParquetDB solves and for whom it is intended.

Thank you again for your input!

Best Regards,

Logan Lang
