### 2025-05-14 lllangWV/ParquetDB#43 @fabian-s
Polishing JOSS paper
Some minor edits:

* "table and field level metadata."  change to **table- and field-level metadata.**

* "Leveraging Apache Parquet(“Parquet,” n.d.; Welcome to The Apache Software Foundation, n.d.)," : insert a space and lowercase the article: "Leveraging Apache Parquet (“Parquet,” n.d.; Welcome to the Apache Software Foundation, n.d.),"

* "While binary formats like HDF5 (?) improve efficiency"  : supply missing citation

* "medium ware" (twice) : use the standard term "middleware"?

* "Pandas(Pandas - Python Data Analysis Library, n.d.), Dask(Dask | Scale the Python Tools You Love, n.d.), and Polars(Polars, n.d.), or embedded analytical databases such as DuckDB(User, n.d.)," : add spaces before each parenthesis and simplify the citations: "Pandas (n.d.), Dask (n.d.), Polars (n.d.), and the embedded analytical database DuckDB (n.d.)"

* "ParquetDB’s create times are competitive"   : use  proper noun "ParquetDB’s creation times are competitive"

* "benefiting from Parquet’s columnar efficiency. (See Figure 1)" : move figure tag inside the sentence and drop the capital: "benefiting from Parquet’s columnar efficiency (see Figure 1)."

* "Detailed benchmarks are in our extended paper ([https://arxiv.org/abs/2502.05311)(](https://arxiv.org/abs/2502.05311%29%28)?)."  fuse  parentheses and remove the "?": "Detailed benchmarks are provided in our extended paper ([https://arxiv.org/abs/2502.05311](https://arxiv.org/abs/2502.05311))."

* "Advanced Cyberinfrastructure Co- ordination Ecosystem"  : remove stray hyphen-space: "Advanced Cyberinfrastructure Coordination Ecosystem"

* "rec-ommendation" (line-broken word) : "recommendation"

* "Practical Issues in Database Management: A **Refernce* for the Thinking Practitioner"  : typo

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7932
