### 2025-05-14 ff241c1
fix(docs): Correct terminology from "medium ware" to "middleware"
Updated the terminology in the paper.md and bibliography files to
accurately reflect the nature of ParquetDB as a middleware solution.
This change aims to enhance clarity and ensure consistency in the
documentation.
