### 2025-05-14 d8c36e1
fix(docs): Update paper.md and bibliography for clarity and accuracy
Revised text in paper.md to improve clarity, including adjustments to
punctuation and phrasing. Added a new entry in the bibliography for
DuckDB, enhancing the documentation's comprehensiveness. These changes
aim to ensure precise communication of ParquetDB's capabilities and
context within the field.
