+++ joss/ParquetDB.bib
@@ -54,6 +54,7 @@
 
 @misc{HDF5PythonH5py,
   title = {{{HDF5}} for {{Python}} --- H5py 3.13.0 Documentation},
+  shorttitle = {{{HDF5}}},
   urldate = {2025-04-28},
   howpublished = {https://docs.h5py.org/en/stable/index.html}
 }
@@ -75,6 +76,16 @@
   langid = {english}
 }
 
+@misc{InprocessSQLOLAP,
+  title = {An In-Process {{SQL OLAP}} Database Management System},
+  shorttitle = {{{DuckDB}}},
+  journal = {DuckDB},
+  urldate = {2025-05-13},
+  abstract = {DuckDB is an in-process SQL OLAP database management system. Simple, feature-rich, fast \& open source.},
+  howpublished = {https://duckdb.org/},
+  langid = {english}
+}
+
 @article{jainCommentaryMaterialsProject2013,
   title = {Commentary: {{The Materials Project}}: {{A}} Materials Genome Approach to Accelerating Materials Innovation},
   shorttitle = {Commentary},
@@ -89,6 +100,24 @@
   doi = {10.1063/1.4812323},
   urldate = {2024-10-20},
   abstract = {Accelerating the discovery of advanced materials is essential for human welfare and sustainable, clean energy. In this paper, we introduce the Materials Project (www.materialsproject.org), a core program of the Materials Genome Initiative that uses high-throughput computing to uncover the properties of all known inorganic materials. This open dataset can be accessed through multiple channels for both interactive exploration and data mining. The Materials Project also seeks to create open-source platforms for developing robust, sophisticated materials analyses. Future efforts will enable users to perform ``rapid-prototyping'' of new materials in silico, and provide researchers with new avenues for cost-effective, data-driven materials design.}
+}
+
+@misc{langParquetDBLightweightPython2025,
+  title = {{{ParquetDB}}: {{A Lightweight Python Parquet-Based Database}}},
+  shorttitle = {{{ParquetDB}}},
+  author = {Lang, Logan and Hernandez, Eduardo and Choudhary, Kamal and Romero, Aldo H.},
+  year = {2025},
+  month = feb,
+  number = {arXiv:2502.05311},
+  eprint = {2502.05311},
+  primaryclass = {cs},
+  publisher = {arXiv},
+  doi = {10.48550/arXiv.2502.05311},
+  urldate = {2025-02-14},
+  abstract = {Traditional data storage formats and databases often introduce complexities and inefficiencies that hinder rapid iteration and adaptability. To address these challenges, we introduce ParquetDB, a Python-based database framework that leverages the Parquet file format's optimized columnar storage. ParquetDB offers efficient serialization and deserialization, native support for complex and nested data types, reduced dependency on indexing through predicate pushdown filtering, and enhanced portability due to its file-based storage system. Benchmarks show that ParquetDB outperforms traditional databases like SQLite and MongoDB in managing large volumes of data, especially when using data formats compatible with PyArrow. We validate ParquetDB's practical utility by applying it to the Alexandria 3D Materials Database, efficiently handling approximately 4.8 million complex and nested records. By addressing the inherent limitations of existing data storage systems and continuously evolving to meet future demands, ParquetDB has the potential to significantly streamline data management processes and accelerate research development in data-driven fields.},
+  archiveprefix = {arXiv},
+  langid = {english},
+  keywords = {Computer Science - Databases,Physics - Data Analysis Statistics and Probability}
 }
 
 @misc{PandasPythonData,
@@ -143,18 +172,9 @@
   langid = {english}
 }
 
[…]
