### 2025-05-14 8fe980f
fix(docs): Correct bibliography entry for Dask and improve text clarity
Updated the short title for the Dask bibliography entry to enhance
readability. Revised phrasing in paper.md to ensure consistent
terminology and improve overall clarity of the document. These
changes aim to provide a more accurate representation of the
tools and context relevant to ParquetDB.
