+++ joss/ParquetDB.bib
@@ -14,6 +14,7 @@
 
 @misc{DaskScalePythona,
   title = {Dask {\textbar} {{Scale}} the {{Python}} Tools You Love},
+  shorttitle = {Dask},
   urldate = {2025-05-13},
   abstract = {Dask is a flexible open-source Python library for parallel computing maintained by OSS contributors across dozens of companies including Anaconda, Coiled, SaturnCloud, and nvidia.},
   howpublished = {https://www.dask.org/},
