### 2025-05-14 6127755
fix(docs): Removing hyphens in acknowledgements
Revised the acknowledgements in paper.md to specify that certain
commercial equipment and materials are identified for procedural
clarity, not as endorsements. This change aims to enhance the
accuracy of the document and prevent potential misinterpretations.
