+++ joss/paper.md
@@ -31,6 +31,9 @@ affiliations:
     name: National Institute of Standards and Technology, Gaithersburg, United States
 date: 14 February 2025
 bibliography: ParquetDB.bib
+header-includes: |
+  \hyphenpenalty=10000
+  \exhyphenpenalty=10000
 ---
 
 # Summary
@@ -69,7 +72,7 @@ For more details, please visit the [GitHub repository](https://github.com/lllang
 
 # Acknowledgements
 
-We thank the Pittsburgh Supercomputer Center (Bridges2) and San Diego Supercomputer Center (Expanse) through allocation DMR140031 from the Advanced Cyberinfrastructure \mbox{Coordination} Ecosystem: Services \& Support (ACCESS) program, which is supported by \mbox{National} Science Foundation grants \#2138259, \#2138286, \#2138307, \#2137603, and \#2138296. We gratefully acknowledge the computational resources provided by the WVU Research \mbox{Computing} Dolly Sods HPC cluster, partially funded by NSF OAC-2117575. Additionally, we recognize the support from the West Virginia Higher Education Policy Commission through the Research Challenge Grant Program 2022 (Award RCG 23-007), as well as NASA EPSCoR (Award 80NSSC22M0173), for their contributions to this work. The work of E.R.H.  is supported by MCIN/AEI/ 10.13039/501100011033/FEDER, UE through projects PID2022-139776NB-C66. K.C. thanks funding from the CHIPS Metrology Program, part of CHIPS for America, National Institute of Standards and Technology, U.S. Department of Commerce.  Certain commercial equipment, instruments, software, or materials are identified in this paper in order to specify the experimental procedure adequately. Such identifications are not intended to imply recommendation or endorsement by NIST, nor are they intended to imply that the materials or equipment identified are necessarily the best available for the purpose.
+We thank the Pittsburgh Supercomputer Center (Bridges2) and San Diego Supercomputer Center (Expanse) through allocation DMR140031 from the Advanced Cyberinfrastructure \mbox{Coordination} Ecosystem: Services \& Support (ACCESS) program, which is supported by \mbox{National} Science Foundation grants \#2138259, \#2138286, \#2138307, \#2137603, and \#2138296. We gratefully acknowledge the computational resources provided by the WVU Research \mbox{Computing} Dolly Sods HPC cluster, partially funded by NSF OAC-2117575. Additionally, we recognize the support from the West Virginia Higher Education Policy Commission through the Research Challenge Grant Program 2022 (Award RCG 23-007), as well as NASA EPSCoR (Award 80NSSC22M0173), for their contributions to this work. The work of E.R.H.  is supported by MCIN/AEI/ 10.13039/501100011033/FEDER, UE through projects PID2022-139776NB-C66. K.C. thanks funding from the CHIPS Metrology Program, part of CHIPS for America, National Institute of Standards and Technology, U.S. Department of Commerce.  Certain commercial equipment, instruments, software, or materials are identified in this paper in order to specify the experimental procedure adequately. Such identifications are not intended to imply \mbox{recommendation} or endorsement by NIST, nor are they intended to imply that the materials or equipment identified are necessarily the best available for the purpose.
 
 # References
 
