### 2025-05-14 eb4c76c
fix(docs): Correct reference typo in bibliography and update link format
Revised the title in the bibliography entry for clarity by correcting
the word "Refernce" to "Reference". Updated the format of the GitHub
repository link in paper.md for consistency. These changes aim to
improve the accuracy and readability of the documentation.
