+++ joss/ParquetDB.bib
@@ -138,7 +138,7 @@
 }
 
 @book{pascalPracticalIssuesDatabase2000,
-  title = {Practical {{Issues}} in {{Database Management}}: {{A Refernce}} for the {{Thinking Practitioner}}},
+  title = {Practical {{Issues}} in {{Database Management}}: {{A Reference}} for the {{Thinking Practitioner}}},
   shorttitle = {Practical {{Issues}} in {{Database Management}}},
   author = {Pascal, Fabian},
   year = {2000},
