### 2025-05-14 lllangWV/ParquetDB#43 @lllangWV
Hey @fabian-s ,

Thanks for reviewing the manuscript and suggesting edits. I have addressed all the edits above in the paper.

Best Regards,

Logan Lang
