### 2025-05-15 811da45
bug(utils): Correct index handling in update_flattend_table function for multi key updates
Updated the index handling logic in the update_flattend_table function
to ensure proper selection and sorting of incoming and current tables.
This change addresses a bug where the updated table would be out of order after the right join. This is because pyarrow does not ensure the row order is preserved.
