### 2025-05-16 lllangWV/ParquetDB#43 @fabian-s
Thanks, one final nitpick

-- the DuckDB reference seems to be malformed:

> DuckDB (“DuckDB,” n.d.)

-- get rid of the quotes, please, and make sure the reference is formatted like the other software references.
