### 2025-05-16 #7932 @fabian-s
One final minor thing to fix (see https://github.com/lllangWV/ParquetDB/issues/43), then please

- Re-check the references (`@editorialbot check references`)
