### 2025-05-16 b6f22bb
fix: Correct shorttitle formatting in ParquetDB.bib
Updated the shorttitle field for the In-Process SQL OLAP Database
Management System entry to remove unnecessary braces. This change
improves consistency and readability in the bibliography.
