+++ joss/ParquetDB.bib
@@ -79,8 +79,7 @@
 
 @misc{InprocessSQLOLAP,
   title = {An In-Process {{SQL OLAP}} Database Management System},
-  shorttitle = {{{DuckDB}}},
-  journal = {DuckDB},
+  shorttitle = {DuckDB},
   urldate = {2025-05-13},
   abstract = {DuckDB is an in-process SQL OLAP database management system. Simple, feature-rich, fast \& open source.},
   howpublished = {https://duckdb.org/},
