### 2025-05-16 lllangWV/ParquetDB#43 @lllangWV
@fabian-s

Okay, I have fixed this.

For some reason zotero with the better bibtex extension was exporting the DuckDB reference with quotations, but it was not doing it with others .

Logan Lang
