### 2025-05-16 1a58161
Update CHANGELOG for v1.0.0
