### 2025-05-16 3ca761d
Update CHANGELOG for v1.0.1
