### 2025-05-16 #7932 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
