+++ joss/ParquetDB.bib
@@ -1,5 +1,5 @@
 @book{allenDefinitiveGuideSQLite2010,
-  title = {The {{Definitive Guide}} to {{SQLite}}},
+  title = {The Definitive Guide to {SQLite}},
   author = {Allen, Grant and Owens, Mike},
   year = {2010},
   publisher = {Apress},
@@ -13,16 +13,16 @@
 }
 
 @misc{DaskScalePythona,
-  title = {Dask {\textbar} {{Scale}} the {{Python}} Tools You Love},
+  title = {Dask {\textbar} {S}cale the {P}ython} Tools You Love},
   shorttitle = {Dask},
   urldate = {2025-05-13},
   abstract = {Dask is a flexible open-source Python library for parallel computing maintained by OSS contributors across dozens of companies including Anaconda, Coiled, SaturnCloud, and nvidia.},
-  howpublished = {https://www.dask.org/},
+  url = {https://www.dask.org/},
   langid = {english}
 }
 
 @article{guoMongoDBsJavaScriptFuzzer2017,
-  title = {{{MongoDB}}'s {{JavaScript}} Fuzzer},
+  title = {{MongoDB}'s {JavaScript} Fuzzer},
   author = {Guo, Robert},
   year = {2017},
   month = apr,
@@ -54,10 +54,10 @@
 }
 
 @misc{HDF5PythonH5py,
-  title = {{{HDF5}} for {{Python}} --- H5py 3.13.0 Documentation},
+  title = {{HDF5} for {P}ython --- H5py 3.13.0 Documentation},
   shorttitle = {{{HDF5}}},
   urldate = {2025-04-28},
-  howpublished = {https://docs.h5py.org/en/stable/index.html}
+  url = {https://docs.h5py.org/en/stable/index.html}
 }
 
 @article{hjorthlarsenAtomicSimulationEnvironment2017,
@@ -82,7 +82,7 @@
   shorttitle = {DuckDB},
   urldate = {2025-05-13},
   abstract = {DuckDB is an in-process SQL OLAP database management system. Simple, feature-rich, fast \& open source.},
-  howpublished = {https://duckdb.org/},
+  url = {https://duckdb.org/},
   langid = {english}
 }
 
@@ -103,7 +103,7 @@
 }
 
 @misc{langParquetDBLightweightPython2025,
-  title = {{{ParquetDB}}: {{A Lightweight Python Parquet-Based Database}}},
+  title = {{ParquetDB}: A Lightweight {P}ython {P}arquet-Based Database},
   shorttitle = {{{ParquetDB}}},
   author = {Lang, Logan and Hernandez, Eduardo and Choudhary, Kamal and Romero, Aldo H.},
   year = {2025},
@@ -121,10 +121,10 @@
 }
 
 @misc{PandasPythonData,
-  title = {Pandas - {{Python Data Analysis Library}}},
+  title = {Pandas - {P}ython Data Analysis Library},
   shorttitle = {Pandas},
   urldate = {2025-05-13},
-  howpublished = {https://pandas.pydata.org/}
+  url = {https://pandas.pydata.org/}
 }
 
 @misc{Parquet,
@@ -132,12 +132,12 @@
   journal = {Apache Parquet},
   urldate = {2024-10-21},
   abstract = {The Apache Parquet Website},
-  howpublished = {https://parquet.apache.org/},
+  url = {https://parquet.apache.org/},
   langid = {english}
 }
 
 @book{pascalPracticalIssuesDatabase2000,
-  title = {Practical {{Issues}} in {{Database Management}}: {{A Reference}} for the {{Thinking Practitioner}}},
+  title = {Practical Issues in Database Management: A Reference for the Thinking Practitioner},
[…]
