### 2025-05-16 1e9f5ab
fix: Correct title formatting in ParquetDB.bib
Adjusted the title formatting in the Dask citation entry to ensure
consistency and proper display in the bibliography.
