+++ joss/ParquetDB.bib
@@ -13,7 +13,7 @@
 }
 
 @misc{DaskScalePythona,
-  title = {Dask {\textbar} {S}cale the {P}ython} Tools You Love},
+  title = {Dask {\textbar} {S}cale the {P}ython Tools You Love},
   shorttitle = {Dask},
   urldate = {2025-05-13},
   abstract = {Dask is a flexible open-source Python library for parallel computing maintained by OSS contributors across dozens of companies including Anaconda, Coiled, SaturnCloud, and nvidia.},
