### 2025-05-16 #7932 @lllangWV
Hey @danielskatz,

Thanks for catching this! I have edited the Zenodo metadata to "1.0.1".

I also merged [lllangWV/ParquetDB#53](https://github.com/lllangWV/ParquetDB/pull/53). There was an issue in the git actions related to and extra curly bracket in `ParquetDB.bib`, but I have fixed this and the paper compiled correctly.

Best Regards,

Logan Lang
