### 2025-05-16 #7932 @danielskatz
Congratulations to @lllangWV (Logan Lang) and co-authors on your JOSS publication!!

JOSS depends on volunteers, and couldn't do this without you
