### 2025-05-16 #7932 @lllangWV
Thank you to @ckoerber and @perdelt for their thorough reviews and valuable feedback on ParquetDB! We really appreciate the time and effort you dedicated to helping us improve the paper and the package.

And thank you to @fabian-s for your editorial guidance throughout the process.

Thank you @danielskatz for the edits and suggestions you provided earlier today!

We're thrilled to see ParquetDB published in JOSS. This wouldn't be possible without your contributions!

Best regards,

Logan Lang
